feat(wallet): unify the create-wallet flow across desktop and mobile
One CreateWalletFlowComponent (features/wallet-setup) now serves both the
desktop /auth wizard and the mobile shell, best-of-both:

- StepHeader progress (4 steps) over the mobile card layout
- shared WalletNameSectionComponent with prefilled suggestion, and the
  name-conflict/invalid messages now actually VISIBLE (they were
  mat-error on a validator-less ngModel, which Material never renders)
- desktop verify style (per-field inline errors, Next gated on all
  correct) and desktop draft hygiene (back-nav abandons the phrase,
  Generate New resets the BIP39 fields)
- mobile protect step (optional at-rest passphrase, confirm + warning
  once typed) with per-mode wording; commit branches Core RPC vs BDK
  via NodeService.isRemote, hosts handle post-create navigation

Supporting unifications along the way:

- MnemonicDisplayComponent is the single phrase grid: CSS-counter
  indexes (numbers no longer copyable -> selecting yields one word per
  line, no clipboard magic), container-query column count (4 only when
  they fit, 3 down to the 320px floor, no viewport queries), white
  chips, Generate New reused on mobile
- new shared Bip39PassphraseSectionComponent (create/restore modes)
  replaces four hand-rolled 25th-word sections; desktop import loses
  its pointless confirm field on restore
- suppress the WebView2/Edge native password-reveal eye globally
  (input::-ms-reveal) - secret fields showed two toggle icons
- i18n: bip39/backup/verify key redundancy collapsed (12->6 bip39 keys,
  mwallet_* duplicates deleted across all 26 locales by copying
  existing translations - zero new translation work)
