+++ web-wallet/src/app/features/auth/pages/create-wallet/create-wallet.component.ts
@@ -1,689 +1,54 @@
-import {
-  Component,
-  inject,
-  signal,
-  viewChildren,
-  ElementRef,
-  OnInit,
-  OnDestroy,
-} from '@angular/core';
-
-import { Router, RouterModule } from '@angular/router';
-import { FormsModule } from '@angular/forms';
-import { MatCardModule } from '@angular/material/card';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIconModule } from '@angular/material/icon';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatCheckboxModule } from '@angular/material/checkbox';
+import { Component, inject } from '@angular/core';
+import { Router } from '@angular/router';
 import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatProgressBarModule } from '@angular/material/progress-bar';
-import { MatAutocompleteModule, MatAutocompleteTrigger } from '@angular/material/autocomplete';
-import { MatTooltipModule } from '@angular/material/tooltip';
-import { toSignal } from '@angular/core/rxjs-interop';
-import { Store } from '@ngrx/store';
-import { I18nPipe, I18nService } from '../../../../core/i18n';
-import { StepHeaderComponent, MnemonicDisplayComponent } from '../../../../shared/components';
-import { WalletManagerService } from '../../../../bitcoin/services/wallet/wallet-manager.service';
-import { DescriptorService } from '../../../../bitcoin/services/wallet/descriptor.service';
-import { selectIsTestnet } from '../../../../store/settings/settings.selectors';
-import { NodeService } from '../../../../node/services/node.service';
-import { BtcxWalletService } from '../../../../core/services/btcx-wallet.service';
+import { I18nService } from '../../../../core/i18n';
+import { CreateWalletFlowComponent } from '../../../wallet-setup/create-wallet-flow.component';
 
 /**
- * CreateWalletComponent guides users through creating a new Bitcoin wallet.
- * Steps:
- * 1. Enter wallet name
- * 2. Generate and display mnemonic + optional BIP39 passphrase (25th word)
- * 3. Verify mnemonic (confirm 3 random words)
- * 4. Optional Bitcoin Core wallet encryption
+ * CreateWalletComponent — the desktop /auth host for the shared
+ * CreateWalletFlowComponent (the wizard itself lives in
+ * features/wallet-setup and is also mounted by the mobile shell).
+ * This host only provides the auth-page chrome and post-create navigation.
  */
 @Component({
   selector: 'app-create-wallet',
   standalone: true,
-  imports: [
-    RouterModule,
-    FormsModule,
-    MatCardModule,
-    MatButtonModule,
-    MatIconModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatCheckboxModule,
-    MatSnackBarModule,
-    MatProgressBarModule,
-    MatAutocompleteModule,
-    MatTooltipModule,
-    I18nPipe,
-    StepHeaderComponent,
-    MnemonicDisplayComponent,
-  ],
+  imports: [MatSnackBarModule, CreateWalletFlowComponent],
   template: `
     <div class="create-wallet-container">
-      <mat-card class="create-card">
-        <app-step-header
-          [title]="getCurrentStepTitle()"
-          [currentStep]="currentStep()"
-          [totalSteps]="4"
-        ></app-step-header>
-
-        @if (creating()) {
-          <mat-progress-bar mode="indeterminate"></mat-progress-bar>
-        }
-
-        <mat-card-content>
-          <!-- Step 1: Wallet Name -->
-          @if (currentStep() === 1) {
-            <div class="step-content">
-              <mat-form-field appearance="outline" class="full-width">
-                <mat-label>{{ 'wallet_name' | i18n }}</mat-label>
-                <input
-                  matInput
-                  [(ngModel)]="walletName"
-                  (ngModelChange)="onWalletNameChange()"
-                  placeholder="My Wallet"
-                  [disabled]="creating()"
-                />
-                @if (walletNameConflict()) {
-                  <mat-error>{{ 'wallet_name_conflict' | i18n }}</mat-error>
-                } @else if (walletNameInvalid()) {
-                  <mat-error>{{ 'wallet_name_invalid_local' | i18n }}</mat-error>
-                } @else {
-                  <mat-hint>{{
-                    (isRemote() ? 'wallet_name_hint_local' : 'wallet_name_hint') | i18n
-                  }}</mat-hint>
-                }
-              </mat-form-field>
-              <div class="step-actions">
-                <button mat-button routerLink="/auth">
-                  {{ 'back' | i18n }}
-                </button>
-                <button
-                  mat-raised-button
-                  color="primary"
-                  [disabled]="
-                    !walletName || walletNameConflict() || walletNameInvalid() || creating()
-                  "
-                  (click)="nextStep()"
-                >
-                  {{ 'next' | i18n }}
-                </button>
-              </div>
-            </div>
-          }
-
-          <!-- Step 2: Generate Mnemonic + Optional BIP39 Passphrase -->
-          @if (currentStep() === 2) {
-            <div class="step-content">
-              <p class="warning-text">
-                <mat-icon>warning</mat-icon>
-                {{ 'backup_warning' | i18n }}
-              </p>
-
-              <app-mnemonic-display
-                [words]="mnemonicWords()"
-                [disabled]="creating()"
-                (regenerate)="generateMnemonic()"
-              ></app-mnemonic-display>
-
-              <!-- BIP39 Passphrase Option (25th word) — supported on BOTH the
-                   Core and the local BDK (remote/mobile) wallet paths. It is
-                   folded into the derivation, so with it set the mnemonic
-                   alone no longer recovers the funds; encryption-at-rest is a
-                   separate choice offered on the last step. -->
-              <div class="passphrase-section">
-                <mat-checkbox [(ngModel)]="useBip39Passphrase" class="passphrase-checkbox">
-                  {{ 'use_bip39_passphrase' | i18n }}
-                </mat-checkbox>
-
-                @if (useBip39Passphrase) {
-                  <p class="info-text small">{{ 'bip39_passphrase_info' | i18n }}</p>
-
-                  <p class="warning-text small">
-                    <mat-icon>warning</mat-icon>
-                    {{ 'bip39_passphrase_warning' | i18n }}
-                  </p>
-
-                  <mat-form-field appearance="outline" class="full-width">
-                    <mat-label>{{ 'bip39_passphrase' | i18n }}</mat-label>
-                    <!-- The 25th word is byte-significant: keep the browser /
-                         Android from auto-capitalizing/autocorrecting it when
-                         revealed. -->
-                    <input
-                      matInput
-                      [type]="bip39Visible() ? 'text' : 'password'"
-                      [(ngModel)]="passphrase"
-                      [disabled]="creating()"
-                      autocomplete="off"
-                      autocapitalize="none"
-                      autocorrect="off"
-                      spellcheck="false"
-                    />
-                    <button
-                      mat-icon-button
-                      matSuffix
-                      type="button"
-                      (click)="bip39Visible.set(!bip39Visible())"
-                      [attr.aria-label]="
-                        (bip39Visible() ? 'hide_passphrase' : 'show_passphrase') | i18n
-                      "
-                      [matTooltip]="(bip39Visible() ? 'hide_passphrase' : 'show_passphrase') | i18n"
-                    >
-                      <mat-icon>{{ bip39Visible() ? 'visibility_off' : 'visibility' }}</mat-icon>
-                    </button>
-                  </mat-form-field>
-
-                  <mat-form-field appearance="outline" class="full-width">
-                    <mat-label>{{ 'confirm_bip39_passphrase' | i18n }}</mat-label>
-                    <input
-                      matInput
-                      [type]="bip39Visible() ? 'text' : 'password'"
-                      [(ngModel)]="passphraseConfirm"
-                      [disabled]="creating()"
-                      autocomplete="off"
-                      autocapitalize="none"
-                      autocorrect="off"
-                      spellcheck="false"
-                    />
-                    @if (passphrase !== passphraseConfirm && passphraseConfirm) {
-                      <mat-error>{{ 'passphrase_mismatch' | i18n }}</mat-error>
-                    }
-                  </mat-form-field>
-                }
-              </div>
-
-              <mat-checkbox [(ngModel)]="mnemonicWrittenDown" class="confirm-checkbox">
-                {{ 'confirm_backup_written' | i18n }}
-              </mat-checkbox>
-
-              <div class="step-actions">
-                <button mat-button (click)="prevStep()" [disabled]="creating()">
-                  {{ 'back' | i18n }}
-                </button>
-                <button
-                  mat-raised-button
-                  color="primary"
-                  [disabled]="!mnemonicWrittenDown || creating() || !bip39PassphraseValid()"
-                  (click)="nextStep()"
-                >
-                  {{ 'next' | i18n }}
-                </button>
-              </div>
-            </div>
-          }
-
-          <!-- Step 3: Verify Mnemonic -->
-          @if (currentStep() === 3) {
-            <div class="step-content">
-              <p class="info-text">{{ 'verify_backup_instruction' | i18n }}</p>
-
-              @for (idx of verifyIndices; track idx; let i = $index) {
-                <mat-form-field appearance="outline" class="full-width">
-                  <mat-label>{{ 'word_number' | i18n: { number: idx + 1 } }}</mat-label>
-                  <input
-                    #wordInput
-                    matInput
-                    [(ngModel)]="verifyWords[i]"
-                    [disabled]="creating()"
-                    [matAutocomplete]="auto"
-                    #trigger="matAutocompleteTrigger"
-                    (input)="updateSuggestions(i, verifyWords[i])"
-                    (keydown.enter)="onWordEnter(i, $event, trigger)"
-                    autocomplete="off"
-                  />
-                  <mat-autocomplete
-                    #auto="matAutocomplete"
-                    [autoActiveFirstOption]="true"
-                    (optionSelected)="onWordSelected(i, $event.option.value)"
-                  >
-                    @for (word of wordSuggestions[i]; track word) {
-                      <mat-option [value]="word">{{ word }}</mat-option>
-                    }
-                  </mat-autocomplete>
-                  @if (verifyWords[i] && verifyWords[i] !== mnemonicWords()[idx]) {
-                    <mat-error>{{ 'incorrect_word' | i18n }}</mat-error>
-                  }
-                </mat-form-field>
-              }
-
-              <div class="step-actions">
-                <button mat-button (click)="prevStep()" [disabled]="creating()">
-                  {{ 'back' | i18n }}
-                </button>
-                <button
-                  mat-raised-button
-                  color="primary"
-                  [disabled]="!verificationPassed() || creating()"
-                  (click)="nextStep()"
-                >
-                  {{ 'next' | i18n }}
-                </button>
-              </div>
-            </div>
-          }
-
-          <!-- Step 4: Wallet Encryption (Core passphrase / local seed-at-rest) -->
-          @if (currentStep() === 4) {
-            <div class="step-content">
-              <p class="info-text">
-                {{
-                  (isRemote() ? 'wallet_encryption_info_local' : 'wallet_encryption_info') | i18n
-                }}
-              </p>
-
-              <mat-checkbox [(ngModel)]="useWalletEncryption" class="encryption-checkbox">
-                {{ 'encrypt_wallet' | i18n }}
-              </mat-checkbox>
-
-              @if (useWalletEncryption) {
-                <p class="warning-text">
-                  <mat-icon>warning</mat-icon>
-                  {{ 'wallet_encryption_warning' | i18n }}
-                </p>
-
-                <mat-form-field appearance="outline" class="full-width">
-                  <mat-label>{{ 'wallet_password' | i18n }}</mat-label>
-                  <input
-                    matInput
-                    type="password"
-                    [(ngModel)]="walletPassword"
-                    [disabled]="creating()"
-                  />
-                </mat-form-field>
-
-                <mat-form-field appearance="outline" class="full-width">
-                  <mat-label>{{ 'confirm_wallet_password' | i18n }}</mat-label>
-                  <input
-                    matInput
-                    type="password"
-                    [(ngModel)]="walletPasswordConfirm"
-                    [disabled]="creating()"
-                  />
-                  @if (walletPassword !== walletPasswordConfirm && walletPasswordConfirm) {
-                    <mat-error>{{ 'password_mismatch' | i18n }}</mat-error>
-                  }
-                </mat-form-field>
-              }
-
-              <div class="step-actions">
-                <button mat-button (click)="prevStep()" [disabled]="creating()">
-                  {{ 'back' | i18n }}
-                </button>
-                <button
-                  mat-raised-button
-                  color="primary"
-                  [disabled]="creating() || !walletEncryptionValid()"
-                  (click)="createWallet()"
-                >
-                  @if (creating()) {
-                    {{ 'creating' | i18n }}...
-                  } @else {
-                    {{ 'create_wallet' | i18n }}
-                  }
-                </button>
-              </div>
-            </div>
-          }
-        </mat-card-content>
-      </mat-card>
+      <app-create-wallet-flow
+        [showCancel]="true"
+        (cancelled)="router.navigate(['/auth'])"
+        (created)="onCreated($event)"
+      />
     </div>
   `,
   styles: [
     `
       .create-wallet-container {
         min-height: 100vh;
-        display: flex;
-        align-items: center;
-        justify-content: center;
         padding: 24px;
         background: #eceff1;
-      }
-
-      .create-card {
-        width: 100%;
-        max-width: 600px;
-      }
-
-      .step-content {
-        padding: 24px;
-      }
-
-      .full-width {
-        width: 100%;
-      }
-
-      .warning-text {
         display: flex;
+        justify-content: center;
         align-items: flex-start;
-        gap: 8px;
-        color: #e65100;
-        background: rgba(255, 152, 0, 0.1);
-        padding: 12px;
-        border-radius: 4px;
-        margin-bottom: 16px;
-        font-size: 13px;
-
-        mat-icon {
-          flex-shrink: 0;
-          color: #ff9800;
-        }
       }
 
-      .info-text {
-        color: rgba(0, 0, 0, 0.6);
-        margin-bottom: 16px;
-      }
-
-      .confirm-checkbox {
-        margin: 16px 0;
-      }
-
-      .passphrase-section {
-        margin: 16px 0;
-        padding: 16px;
-        background: rgba(0, 0, 0, 0.02);
-        border-radius: 4px;
-        border: 1px solid rgba(0, 0, 0, 0.08);
-      }
-
-      .passphrase-checkbox,
-      .encryption-checkbox {
-        display: block;
-        margin-bottom: 12px;
-      }
-
-      .info-text.small,
-      .warning-text.small {
-        font-size: 12px;
-        margin-bottom: 12px;
-      }
-
-      .step-actions {
-        display: flex;
-        gap: 8px;
-        justify-content: flex-end;
-        margin-top: 24px;
-        padding-top: 16px;
-        border-top: 1px solid rgba(0, 0, 0, 0.08);
+      app-create-wallet-flow {
+        flex: 1;
       }
     `,
   ],
 })
-export class CreateWalletComponent implements OnInit, OnDestroy {
-  private readonly router = inject(Router);
-  private readonly store = inject(Store);
-  private readonly walletManager = inject(WalletManagerService);
-  private readonly descriptorService = inject(DescriptorService);
+export class CreateWalletComponent {
+  protected readonly router = inject(Router);
   private readonly snackBar = inject(MatSnackBar);
   private readonly i18n = inject(I18nService);
-  private readonly nodeService = inject(NodeService);
-  private readonly btcxWallet = inject(BtcxWalletService);
 
-  /** Remote (Electrum) mode: local BDK wallet, seed handled in Rust. */
-  readonly isRemote = this.nodeService.isRemote;
-
-  // Get testnet flag from settings store
-  private readonly isTestnet = toSignal(this.store.select(selectIsTestnet), { initialValue: true });
-
-  // Step management
-  currentStep = signal(1);
-  private readonly stepTitles = [
-    'wallet_name',
-    'backup_phrase',
-    'verify_backup',
-    'wallet_encryption_title',
-  ];
-
-  walletName = '';
-  mnemonicWrittenDown = false;
-  creating = signal(false);
-
-  // Existing wallet names (for conflict check on step 1)
-  private readonly existingWalletNames = signal<string[]>([]);
-  readonly walletNameConflict = signal(false);
-  mnemonic = '';
-  mnemonicWords = signal<string[]>([]);
-
-  // BIP39 Passphrase (25th word)
-  useBip39Passphrase = false;
-  passphrase = '';
-  passphraseConfirm = '';
-  /** Reveal toggle for the BIP39 25th-word inputs (verify what was typed). */
-  readonly bip39Visible = signal(false);
-
-  // Bitcoin Core Wallet Encryption
-  useWalletEncryption = false;
-  walletPassword = '';
-  walletPasswordConfirm = '';
-
-  // Verification
-  verifyIndices: number[] = [];
-  verifyWords: string[] = ['', '', ''];
-  wordSuggestions: string[][] = [[], [], []];
-  readonly wordInputs = viewChildren<ElementRef<HTMLInputElement>>('wordInput');
-
-  ngOnInit(): void {
-    this.generateMnemonic();
-    this.walletManager
-      .listAllWallets()
-      .then(names => this.existingWalletNames.set(names))
-      // RPC unreachable — skip the check; commit-time RPC will surface the real error.
-      .catch(() => undefined);
-  }
-
-  onWalletNameChange(): void {
-    const target = this.walletName.trim().toLowerCase();
-    this.walletNameConflict.set(
-      target.length > 0 && this.existingWalletNames().some(n => n.toLowerCase() === target)
-    );
-    // Local wallets: mirror the Rust-side rules on the name that actually
-    // gets sent (case is preserved, Core-style; uniqueness is checked
-    // case-insensitively above), so violations show before the commit step.
-    const raw = this.walletName.trim();
-    this.walletNameInvalid.set(
-      this.isRemote() && raw.length > 0 && !/^[A-Za-z0-9_-]{1,32}$/.test(raw)
-    );
-  }
-
-  /** Remote mode: the name fails the local store's naming rules. */
-  readonly walletNameInvalid = signal(false);
-
-  getCurrentStepTitle(): string {
-    return this.i18n.get(this.stepTitles[this.currentStep() - 1]);
-  }
-
-  nextStep(): void {
-    if (this.currentStep() < 4) {
-      // Re-generate mnemonic on re-entry to step 2 after it was cleared by a prior back-nav
-      if (this.currentStep() === 1 && !this.mnemonic) {
-        this.generateMnemonic();
-      }
-      // Generate verification indices when moving to step 3
-      if (this.currentStep() === 2) {
-        this.selectVerificationIndices();
-      }
-      this.currentStep.update(s => s + 1);
-    }
-  }
-
-  prevStep(): void {
-    if (this.currentStep() > 1) {
-      const leaving = this.currentStep();
-      this.currentStep.update(s => s - 1);
-      // User abandons the current mnemonic draft on 2 → 1; a fresh one is generated on re-entry.
-      // 3 → 2 intentionally preserves mnemonic so the user can view it again.
-      if (leaving === 2) {
-        this.clearMnemonicDraft();
-      }
-    }
-  }
-
-  generateMnemonic(): void {
-    this.mnemonic = this.walletManager.generateMnemonic(256);
-    this.mnemonicWords.set(this.mnemonic.split(' '));
-    this.mnemonicWrittenDown = false;
-    this.verifyWords = ['', '', ''];
-    // Reset BIP39 passphrase when generating new mnemonic
-    this.useBip39Passphrase = false;
-    this.passphrase = '';
-    this.passphraseConfirm = '';
-  }
-
-  private selectVerificationIndices(): void {
-    // Select 3 random indices for verification
-    const indices: number[] = [];
-    while (indices.length < 3) {
-      const idx = Math.floor(Math.random() * 24);
-      if (!indices.includes(idx)) {
-        indices.push(idx);
-      }
-    }
-    this.verifyIndices = indices.sort((a, b) => a - b);
-    this.verifyWords = ['', '', ''];
-  }
-
-  verificationPassed(): boolean {
-    return this.verifyIndices.every(
-      (idx, i) =>
-        this.verifyWords[i].toLowerCase().trim() === this.mnemonicWords()[idx].toLowerCase()
-    );
-  }
-
-  updateSuggestions(index: number, value: string): void {
-    if (value && value.length >= 1) {
-      this.wordSuggestions[index] = this.descriptorService.getWordSuggestions(value, 8);
-    } else {
-      this.wordSuggestions[index] = [];
-    }
-  }
-
-  onWordSelected(index: number, word: string): void {
-    this.verifyWords[index] = word;
-    this.wordSuggestions[index] = [];
-    this.focusNext(index);
-  }
-
-  /**
-   * Enter-to-accept, mirroring the import flow's mnemonic entry. When the
-   * autocomplete panel is open with a highlighted option, Material selects it
-   * (→ onWordSelected). Otherwise, if the typed value is an exact BIP39 word,
-   * commit it and advance to the next verification field.
-   */
-  onWordEnter(index: number, event: Event, trigger: MatAutocompleteTrigger): void {
-    if (event.defaultPrevented) return;
-    if (trigger.panelOpen && trigger.activeOption) return;
-    const typed = (this.verifyWords[index] ?? '').toLowerCase().trim();
-    if (this.descriptorService.getWordlist().includes(typed)) {
-      event.preventDefault();
-      this.verifyWords[index] = typed;
-      this.wordSuggestions[index] = [];
-      this.focusNext(index);
-    }
-  }
-
-  private focusNext(index: number): void {
-    const next = this.wordInputs()[index + 1];
-    if (next) {
-      setTimeout(() => {
-        next.nativeElement.focus();
-        next.nativeElement.select();
-      });
-    }
-  }
-
-  bip39PassphraseValid(): boolean {
-    // Valid if not using passphrase, or if using and both match
-    if (!this.useBip39Passphrase) return true;
-    return this.passphrase === this.passphraseConfirm;
-  }
-
-  walletEncryptionValid(): boolean {
-    // Valid if not using encryption, or if using and passwords match (and not empty)
-    if (!this.useWalletEncryption) return true;
-    return this.walletPassword.length > 0 && this.walletPassword === this.walletPasswordConfirm;
-  }
-
-  async createWallet(): Promise<void> {
-    if (this.creating()) return;
-
-    this.creating.set(true);
-
-    try {
-      if (this.isRemote()) {
-        // Remote (Electrum) mode: the local BDK wallet derives BIP-84 at the
-        // POCX coin type in Rust. Two INDEPENDENT optional secrets: the
-        // at-rest encryption password (locks the seed) and the BIP39 25th
-        // word (changes the derivation — with it set the mnemonic alone no
-        // longer recovers the funds).
-        const passphrase =
-          this.useWalletEncryption && this.walletPassword ? this.walletPassword : undefined;
-        const bip39Passphrase = this.useBip39Passphrase ? this.passphrase : undefined;
-        const name = this.walletName.trim();
-        await this.btcxWallet.create(this.mnemonic, passphrase, name, undefined, bip39Passphrase);
-        const loaded = await this.walletManager.refreshLoadedWallets();
-        if (loaded.includes(name)) {
-          this.walletManager.setActiveWallet(name);
-        }
-      } else {
-        // Only use BIP39 passphrase if checkbox is enabled
-        const mnemonicPassphrase = this.useBip39Passphrase ? this.passphrase : undefined;
-
-        const result = await this.walletManager.createWalletFromMnemonic({
-          walletName: this.walletName,
-          mnemonic: this.mnemonic,
-          mnemonicPassphrase,
-          isTestnet: this.isTestnet(),
-          rescan: false,
-        });
-
-        if (!result.success) {
-          throw new Error(result.errors?.join(', ') || 'Import failed');
-        }
-
-        // Encrypt wallet if requested
-        if (this.useWalletEncryption && this.walletPassword) {
-          await this.walletManager.encryptWallet(this.walletName, this.walletPassword);
-        }
-      }
-
-      this.clearSecrets();
-
-      this.snackBar.open(
-        this.i18n.get('wallet_created_success', { name: this.walletName }),
-        undefined,
-        { duration: 3000 }
-      );
-      this.router.navigate(['/dashboard']);
-    } catch (error) {
-      console.error('Failed to create wallet:', error);
-      // Tauri command rejections are plain strings, not Error instances —
-      // surface them instead of a generic message.
-      const errorMessage =
-        error instanceof Error ? error.message : String(error) || 'Failed to create wallet';
-      this.snackBar.open(errorMessage, this.i18n.get('dismiss'), { duration: 5000 });
-      this.creating.set(false);
-    }
-  }
-
-  ngOnDestroy(): void {
-    this.clearSecrets();
-  }
-
-  private clearMnemonicDraft(): void {
-    this.mnemonic = '';
-    this.mnemonicWords.set([]);
-    this.mnemonicWrittenDown = false;
-    this.passphrase = '';
-    this.passphraseConfirm = '';
-    this.useBip39Passphrase = false;
-  }
-
-  private clearSecrets(): void {
-    this.clearMnemonicDraft();
-    this.walletPassword = '';
-    this.walletPasswordConfirm = '';
-    this.useWalletEncryption = false;
-    this.verifyWords = ['', '', ''];
-    this.wordSuggestions = [[], [], []];
+  onCreated(name: string): void {
+    this.snackBar.open(this.i18n.get('wallet_created_success', { name }), undefined, {
+      duration: 3000,
+    });
+    void this.router.navigate(['/dashboard']);
   }
 }
