+++ web-wallet/src/app/features/mobile-wallet/pages/restore/wallet-restore.component.ts
@@ -7,9 +7,11 @@ import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { MatTooltipModule } from '@angular/material/tooltip';
 import { I18nPipe, I18nService } from '../../../../core/i18n';
-import { MnemonicEntryComponent } from '../../../../shared/components';
+import {
+  MnemonicEntryComponent,
+  Bip39PassphraseSectionComponent,
+} from '../../../../shared/components';
 import type { MnemonicEntryState } from '../../../../shared/components';
 import {
   BtcxWalletService,
@@ -62,8 +64,8 @@ import { WalletNameSectionComponent } from '../../components/wallet-name-section
     MatFormFieldModule,
     MatInputModule,
     MatProgressSpinnerModule,
-    MatTooltipModule,
     MnemonicEntryComponent,
+    Bip39PassphraseSectionComponent,
     I18nPipe,
     PageHeaderComponent,
     WalletNameSectionComponent,
@@ -151,42 +153,12 @@ import { WalletNameSectionComponent } from '../../components/wallet-name-section
                  it reads as part of the phrase (matching the create flow).
                  SEPARATE from the at-rest passphrase below. It must match what
                  the seed was created with, or the probe finds no history. -->
-            <mat-checkbox [(ngModel)]="useBip39" [disabled]="restoring()" class="bip39-toggle">
-              {{ 'mwallet_bip39_toggle' | i18n }}
-            </mat-checkbox>
-
-            @if (useBip39) {
-              <p class="warning-text-inline">
-                <mat-icon class="notice-icon">warning</mat-icon>
-                <span>{{ 'mwallet_bip39_restore_warning' | i18n }}</span>
-              </p>
-              <mat-form-field appearance="outline" class="full-width">
-                <mat-label>{{ 'mwallet_bip39_label' | i18n }}</mat-label>
-                <!-- The 25th word is byte-significant: keep Android from
-                     auto-capitalizing/autocorrecting it when revealed. -->
-                <input
-                  matInput
-                  [type]="bip39Visible() ? 'text' : 'password'"
-                  [(ngModel)]="bip39Passphrase"
-                  autocomplete="off"
-                  autocapitalize="none"
-                  autocorrect="off"
-                  spellcheck="false"
-                />
-                <button
-                  mat-icon-button
-                  matSuffix
-                  type="button"
-                  (click)="bip39Visible.set(!bip39Visible())"
-                  [attr.aria-label]="
-                    (bip39Visible() ? 'hide_passphrase' : 'show_passphrase') | i18n
-                  "
-                  [matTooltip]="(bip39Visible() ? 'hide_passphrase' : 'show_passphrase') | i18n"
-                >
-                  <mat-icon>{{ bip39Visible() ? 'visibility_off' : 'visibility' }}</mat-icon>
-                </button>
-              </mat-form-field>
-            }
+            <app-bip39-passphrase-section
+              mode="restore"
+              [disabled]="restoring()"
+              [(enabled)]="useBip39"
+              [(passphrase)]="bip39Passphrase"
+            />
 
             <mat-form-field appearance="outline" class="full-width">
               <mat-label>{{ 'mwallet_passphrase_optional' | i18n }}</mat-label>
@@ -285,30 +257,6 @@ import { WalletNameSectionComponent } from '../../components/wallet-name-section
         width: 100%;
       }
 
-      .bip39-toggle {
-        display: block;
-        margin: 4px 0 8px;
-      }
-
-      .warning-text-inline {
-        display: flex;
-        gap: 8px;
-        align-items: flex-start;
-        background: rgba(230, 81, 0, 0.08);
-        border-radius: 6px;
-        padding: 10px 12px;
-        font-size: 12px;
-        margin: 0 0 12px;
-
-        .notice-icon {
-          color: #e65100;
-          font-size: 18px;
-          width: 18px;
-          height: 18px;
-          flex-shrink: 0;
-        }
-      }
-
       .success-card {
         text-align: center;
 
@@ -355,8 +303,6 @@ export class WalletRestoreComponent implements OnInit {
    */
   useBip39 = false;
   bip39Passphrase = '';
-  /** Reveal toggle for the BIP39 25th-word input (verify what was typed). */
-  readonly bip39Visible = signal(false);
 
   /** Wallet name — pre-filled with the next free default; desktop rules. */
   walletName = '';
