+++ web-wallet/src/app/shared/components/index.ts
@@ -40,4 +40,5 @@ export { StepHeaderComponent } from './step-header/step-header.component';
 // Seed phrase (shared by create-wallet / import-wallet / multisig wizards)
 export { MnemonicDisplayComponent } from './mnemonic-display/mnemonic-display.component';
 export { MnemonicEntryComponent } from './mnemonic-entry/mnemonic-entry.component';
+export { Bip39PassphraseSectionComponent } from './bip39-passphrase-section/bip39-passphrase-section.component';
 export type { MnemonicEntryState } from './mnemonic-entry/mnemonic-entry.component';
