+++ web-wallet/src/app/shared/components/mnemonic-display/mnemonic-display.component.ts
@@ -6,7 +6,7 @@ import { I18nPipe } from '../../../core/i18n';
 /**
  * MnemonicDisplayComponent renders a generated seed phrase as the standard
  * numbered word-chip grid, with an optional "Generate new" action.
- * Shared by the create-wallet and multisig wizards.
+ * Shared by the create-wallet (desktop + mobile) and multisig wizards.
  */
 @Component({
   selector: 'app-mnemonic-display',
@@ -16,7 +16,6 @@ import { I18nPipe } from '../../../core/i18n';
     <div class="mnemonic-display">
       @for (word of words(); track $index) {
         <div class="word-chip">
-          <span class="word-index">{{ $index + 1 }}</span>
           <span class="word-text">{{ word }}</span>
         </div>
       }
@@ -33,13 +32,28 @@ import { I18nPipe } from '../../../core/i18n';
   `,
   styles: [
     `
-      @use 'breakpoints' as bp;
+      /* The host is the size container: column count derives from the
+         actual space the parent card gives us, never from the viewport,
+         so the grid behaves the same in every wizard. */
+      :host {
+        display: block;
+        container-type: inline-size;
+      }
 
       .mnemonic-display {
         display: grid;
-        grid-template-columns: repeat(4, 1fr);
+        /* 3 columns always (Johnny: 3 work even at 320px) — 4 only when
+           they genuinely fit (>=504px: 4 worst-case chips + gaps). */
+        grid-template-columns: repeat(3, 1fr);
         gap: 8px;
         margin-bottom: 16px;
+        counter-reset: seed-word;
+      }
+
+      @container (min-width: 504px) {
+        .mnemonic-display {
+          grid-template-columns: repeat(4, 1fr);
+        }
       }
 
       .word-chip {
@@ -50,12 +64,16 @@ import { I18nPipe } from '../../../core/i18n';
         background: #ffffff;
         border-radius: 4px;
         font-family: 'Roboto Mono', monospace;
-      }
 
-      .word-index {
-        color: rgba(0, 0, 0, 0.4);
-        font-size: 12px;
-        min-width: 20px;
+        /* Index as a pseudo-element: not in the DOM, so selecting the grid
+           copies only the words — one per line, no numbers. */
+        &::before {
+          counter-increment: seed-word;
+          content: counter(seed-word);
+          color: rgba(0, 0, 0, 0.4);
+          font-size: 12px;
+          min-width: 20px;
+        }
       }
 
       .word-text {
@@ -68,15 +86,39 @@ import { I18nPipe } from '../../../core/i18n';
         margin-bottom: 16px;
       }
 
-      @include bp.phone {
+      /* Narrow container (sub-phone cards): tighten chips so the 8-letter
+         worst case (e.g. "category") still fits 3 columns at a 276px
+         container (= 320px viewport in the mobile create page). Regular
+         chips need ~375px for 3 columns — below that, shrink. */
+      @container (max-width: 380px) {
         .mnemonic-display {
-          grid-template-columns: repeat(3, 1fr);
+          gap: 6px 4px;
+        }
+
+        .word-chip {
+          padding: 5px;
+          gap: 4px;
+
+          &::before {
+            min-width: 12px;
+            font-size: 11px;
+          }
+
+          .word-text {
+            /* 12px: an 8-letter word (widest possible, e.g. "category")
+               must fit the 3-column chip at 320px viewport with zero clip. */
+            font-size: 12px;
+          }
         }
       }
 
-      @media (max-width: 400px) {
-        .mnemonic-display {
-          grid-template-columns: repeat(2, 1fr);
+      :host-context(.dark-theme) {
+        .word-chip {
+          background: #333;
+
+          &::before {
+            color: rgba(255, 255, 255, 0.45);
+          }
         }
       }
     `,
