+++ web-wallet/src/assets/locales/bg.json
@@ -73,8 +73,7 @@
   "balance_available": "Наличен баланс",
   "balance_history": "История на баланса",
   "base_target": "Базова цел",
-  "bip39_passphrase": "BIP39 парола",
-  "bip39_passphrase_info": "Можете да добавите опционална парола (понякога наричана 25-та дума) за допълнителна сигурност. Това създава напълно различен портфейл.",
+  "bip39_passphrase": "BIP39 парола (25-та дума)",
   "bip39_passphrase_warning": "Ако зададете парола, ТРЯБВА да я запомните. Без паролата средствата ви не могат да бъдат възстановени дори с 24-те думи!",
   "bitcoin_pocx_network": "Мрежа Bitcoin PoCX",
   "block": "Блок",
@@ -137,7 +136,7 @@
   "compression": "Компресия",
   "confirm": "Потвърдете",
   "confirm_backup_written": "Записах фразата за възстановяване и я съхраних на сигурно място",
-  "confirm_bip39_passphrase": "Потвърдете паролата",
+  "confirm_bip39_passphrase": "Потвърдете 25-ата дума",
   "confirm_send": "Потвърди и изпрати",
   "confirm_send_message": "Моля, прегледайте детайлите на транзакцията преди изпращане.",
   "confirm_transaction": "Потвърдете транзакцията",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Портфейлът е създаден. Всеки съподписващ трябва да създаде своя портфейл със същите ключове преди получаване на средства.",
   "mwallet_address_failed": "Неуспешно получаване на адрес",
   "mwallet_add_to_contact": "Добави в контактите",
-  "mwallet_backup_ack": "Записах 24-те думи",
-  "mwallet_backup_title": "Запишете фразата си за възстановяване",
-  "mwallet_backup_warning": "Тези 24 думи са единственото резервно копие на портфейла ви. Запишете ги на хартия и ги пазете офлайн. Всеки, който ги знае, може да похарчи средствата ви.",
   "mwallet_branch_legacy": "наследен (coin type 0')",
   "mwallet_create_failed": "Портфейлът не можа да бъде създаден",
   "mwallet_create_wallet": "Създай нов портфейл",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Създайте нов портфейл или възстановете съществуващ от фразата за възстановяване.",
   "mwallet_onboarding_title": "Настройте портфейла си",
   "mwallet_passphrase_confirm": "Потвърди паролата",
-  "mwallet_bip39_toggle": "Добавяне на BIP39 парола (25-та дума)",
-  "mwallet_bip39_label": "BIP39 парола (25-та дума)",
-  "mwallet_bip39_confirm": "Потвърдете 25-ата дума",
-  "mwallet_bip39_warning": "Тази незадължителна 25-та дума се вгражда в ключовете ви. Ако я зададете, фразата за възстановяване САМА по себе си НЯМА да възстанови средствата ви — трябва да помните и точно тази дума. Тя не е паролата на устройството по-горе.",
-  "mwallet_bip39_restore_warning": "Въведете точната 25-та дума, с която е създаден този портфейл. Грешна или липсваща дума извежда друг, празен портфейл.",
+  "bip39_restore_warning": "Въведете точната 25-та дума, с която е създаден този портфейл. Грешна или липсваща дума извежда друг, празен портфейл.",
   "mwallet_passphrase_hint": "Шифрова seed-а, съхранен на това устройство. Тя не е част от фразата за възстановяване — самите думи винаги възстановяват средствата ви.",
   "mwallet_passphrase_label": "Парола",
-  "mwallet_passphrase_mismatch": "Паролите не съвпадат",
   "mwallet_passphrase_optional": "Парола (по избор)",
-  "mwallet_passphrase_title": "Парола на устройството (по избор)",
   "mwallet_received": "Получено",
   "mwallet_remove_server": "Премахни сървъра",
   "mwallet_restore_branch": "Клон на деривация: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Налични",
   "mwallet_title": "Портфейл",
   "mwallet_unlock_failed": "Неуспешно отключване — грешна парола?",
-  "mwallet_verify_hint": "Въведете поисканите думи от записаното копие.",
-  "mwallet_verify_mismatch": "Думите не съвпадат с фразата ви за възстановяване.",
-  "mwallet_verify_title": "Потвърдете резервното копие",
   "mwallet_import_bare_key": "Поставете пълен дескриптор (wpkh(...) или tr(...)), а не самостоятелен ключ.",
   "mwallet_import_bare_wif": "Това е гол частен ключ WIF — обвийте го като wpkh(ВАШИЯТ_WIF), за да го импортирате като портфейл с един адрес.",
   "mwallet_import_description": "Импортирайте портфейл от неговите дескриптори. Поставете един или два частни дескриптора — публичните (xpub) дескриптори са само за наблюдение и още не се поддържат.",
