+++ web-wallet/src/assets/locales/ca.json
@@ -73,8 +73,7 @@
   "balance_available": "Saldo disponible",
   "balance_history": "Historial de saldo",
   "base_target": "Objectiu base",
-  "bip39_passphrase": "Frase de contrasenya BIP39",
-  "bip39_passphrase_info": "Podeu afegir una frase de contrasenya opcional (de vegades anomenada la paraula 25) per a més seguretat. Això crea una cartera completament diferent.",
+  "bip39_passphrase": "Frase de contrasenya BIP39 (25a paraula)",
   "bip39_passphrase_warning": "Si establiu una frase de contrasenya, HEU de recordar-la. Sense la frase de contrasenya, els vostres fons no es poden recuperar ni amb les 24 paraules!",
   "bitcoin_pocx_network": "Xarxa Bitcoin PoCX",
   "block": "Bloc",
@@ -137,7 +136,7 @@
   "compression": "Compressió",
   "confirm": "Confirmar",
   "confirm_backup_written": "He escrit la meva frase de recuperació i l'he guardat de forma segura",
-  "confirm_bip39_passphrase": "Confirmar frase de contrasenya",
+  "confirm_bip39_passphrase": "Confirma la 25a paraula",
   "confirm_send": "Confirmar i enviar",
   "confirm_send_message": "Si us plau, reviseu els detalls de la transacció abans d'enviar.",
   "confirm_transaction": "Confirmar transacció",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Cartera creada. Cada cosignatari ha de crear la seva cartera amb les mateixes claus abans de rebre fons.",
   "mwallet_address_failed": "No s'ha pogut obtenir una adreça",
   "mwallet_add_to_contact": "Afegeix als contactes",
-  "mwallet_backup_ack": "He anotat les 24 paraules",
-  "mwallet_backup_title": "Anoteu la vostra frase de recuperació",
-  "mwallet_backup_warning": "Aquestes 24 paraules són l'única còpia de seguretat de la vostra cartera. Anoteu-les en paper i guardeu-les fora de línia. Qualsevol que les conegui pot gastar els vostres fons.",
   "mwallet_branch_legacy": "antiga (coin type 0')",
   "mwallet_create_failed": "No s'ha pogut crear la cartera",
   "mwallet_create_wallet": "Crear cartera nova",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Creeu una cartera nova o restaureu-ne una d'existent a partir de la frase de recuperació.",
   "mwallet_onboarding_title": "Configureu la vostra cartera",
   "mwallet_passphrase_confirm": "Confirmar contrasenya",
-  "mwallet_bip39_toggle": "Afegeix frase de contrasenya BIP39 (25a paraula)",
-  "mwallet_bip39_label": "Frase de contrasenya BIP39 (25a paraula)",
-  "mwallet_bip39_confirm": "Confirma la 25a paraula",
-  "mwallet_bip39_warning": "Aquesta 25a paraula opcional s'incorpora a les teves claus. Si la configures, la frase de recuperació SOLA NO restaurarà els teus fons — també has de recordar aquesta paraula exacta. No és la contrasenya del dispositiu de dalt.",
-  "mwallet_bip39_restore_warning": "Introdueix la 25a paraula exacta amb què es va crear aquest moneder. Una paraula errònia o absent deriva un moneder diferent i buit.",
+  "bip39_restore_warning": "Introdueix la 25a paraula exacta amb què es va crear aquest moneder. Una paraula errònia o absent deriva un moneder diferent i buit.",
   "mwallet_passphrase_hint": "Xifra la llavor emmagatzemada en aquest dispositiu. No forma part de la frase de recuperació — les paraules per si soles sempre restauren els vostres fons.",
   "mwallet_passphrase_label": "Contrasenya",
-  "mwallet_passphrase_mismatch": "Les contrasenyes no coincideixen",
   "mwallet_passphrase_optional": "Contrasenya (opcional)",
-  "mwallet_passphrase_title": "Contrasenya del dispositiu (opcional)",
   "mwallet_received": "Rebut",
   "mwallet_remove_server": "Eliminar servidor",
   "mwallet_restore_branch": "Branca de derivació: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Disponible",
   "mwallet_title": "Cartera",
   "mwallet_unlock_failed": "No s'ha pogut desbloquejar — contrasenya incorrecta?",
-  "mwallet_verify_hint": "Introduïu les paraules sol·licitades des de la còpia escrita.",
-  "mwallet_verify_mismatch": "Les paraules no coincideixen amb la frase de recuperació.",
-  "mwallet_verify_title": "Confirmeu la còpia de seguretat",
   "mwallet_import_bare_key": "Enganxeu un descriptor complet (wpkh(...) o tr(...)), no una clau sola.",
   "mwallet_import_bare_wif": "Això és una clau privada WIF sola — emboliqueu-la com a wpkh(LA_VOSTRA_WIF) per importar-la com a cartera d'adreça única.",
   "mwallet_import_description": "Importeu una cartera a partir dels seus descriptors. Enganxeu un o dos descriptors privats — els descriptors públics (xpub) són només d'observació i encara no s'admeten.",
