+++ web-wallet/src/assets/locales/cs.json
@@ -73,8 +73,7 @@
   "balance_available": "Dostupny zustatek",
   "balance_history": "Historie zustatku",
   "base_target": "Zakladni cil",
-  "bip39_passphrase": "BIP39 heslo",
-  "bip39_passphrase_info": "Muzete pridat volitelne heslo (nekdy nazyvane 25. slovo) pro dodatecne zabezpeceni. Vytvori to zcela jinou penezenku.",
+  "bip39_passphrase": "BIP39 heslo (25. slovo)",
   "bip39_passphrase_warning": "Pokud nastavite heslo, MUSITE si jej pamatovat. Bez hesla nelze prostredky obnovit ani s 24 slovy!",
   "bitcoin_pocx_network": "Sit Bitcoin PoCX",
   "block": "Blok",
@@ -137,7 +136,7 @@
   "compression": "Komprese",
   "confirm": "Potvrdit",
   "confirm_backup_written": "Zapsal jsem svou obnovovaci frazi a bezpecne ji ulozil",
-  "confirm_bip39_passphrase": "Potvrdit heslo",
+  "confirm_bip39_passphrase": "Potvrdit 25. slovo",
   "confirm_send": "Potvrdit a odeslat",
   "confirm_send_message": "Zkontrolujte podrobnosti transakce pred odeslanim.",
   "confirm_transaction": "Potvrdit transakci",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Peněženka vytvořena. Každý spolupodepisující musí vytvořit svou peněženku se stejnými klíči, než začne přijímat prostředky.",
   "mwallet_address_failed": "Nepodařilo se získat adresu",
   "mwallet_add_to_contact": "Přidat do kontaktů",
-  "mwallet_backup_ack": "Zapsal jsem si 24 slov",
-  "mwallet_backup_title": "Zapište si svou obnovovací frázi",
-  "mwallet_backup_warning": "Těchto 24 slov je jedinou zálohou vaší peněženky. Zapište si je na papír a uchovávejte offline. Kdokoli je zná, může utratit vaše prostředky.",
   "mwallet_branch_legacy": "starší (coin type 0')",
   "mwallet_create_failed": "Peněženku se nepodařilo vytvořit",
   "mwallet_create_wallet": "Vytvořit novou peněženku",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Vytvořte novou peněženku nebo obnovte existující z obnovovací fráze.",
   "mwallet_onboarding_title": "Nastavte si peněženku",
   "mwallet_passphrase_confirm": "Potvrdit heslo",
-  "mwallet_bip39_toggle": "Přidat BIP39 heslo (25. slovo)",
-  "mwallet_bip39_label": "BIP39 heslo (25. slovo)",
-  "mwallet_bip39_confirm": "Potvrdit 25. slovo",
-  "mwallet_bip39_warning": "Toto volitelné 25. slovo se stává součástí vašich klíčů. Pokud jej nastavíte, obnovovací fráze SAMA vaše prostředky NEOBNOVÍ — musíte si pamatovat i toto přesné slovo. Není to heslo zařízení výše.",
-  "mwallet_bip39_restore_warning": "Zadejte přesné 25. slovo, se kterým byla tato peněženka vytvořena. Špatné nebo chybějící slovo odvodí jinou, prázdnou peněženku.",
+  "bip39_restore_warning": "Zadejte přesné 25. slovo, se kterým byla tato peněženka vytvořena. Špatné nebo chybějící slovo odvodí jinou, prázdnou peněženku.",
   "mwallet_passphrase_hint": "Šifruje seed uložený na tomto zařízení. Není součástí obnovovací fráze — samotná slova vždy obnoví vaše prostředky.",
   "mwallet_passphrase_label": "Heslo",
-  "mwallet_passphrase_mismatch": "Hesla se neshodují",
   "mwallet_passphrase_optional": "Heslo (volitelné)",
-  "mwallet_passphrase_title": "Heslo zařízení (volitelné)",
   "mwallet_received": "Přijato",
   "mwallet_remove_server": "Odebrat server",
   "mwallet_restore_branch": "Derivační větev: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "K dispozici",
   "mwallet_title": "Peněženka",
   "mwallet_unlock_failed": "Odemknutí se nezdařilo — špatné heslo?",
-  "mwallet_verify_hint": "Zadejte požadovaná slova ze své písemné zálohy.",
-  "mwallet_verify_mismatch": "Slova neodpovídají vaší obnovovací frázi.",
-  "mwallet_verify_title": "Potvrďte svou zálohu",
   "mwallet_import_bare_key": "Vložte úplný deskriptor (wpkh(...) nebo tr(...)), ne samotný klíč.",
   "mwallet_import_bare_wif": "To je samotný privátní klíč WIF — obalte ho jako wpkh(VÁŠ_WIF) a naimportujte ho jako peněženku s jednou adresou.",
   "mwallet_import_description": "Importujte peněženku z jejích deskriptorů. Vložte jeden nebo dva soukromé deskriptory — veřejné (xpub) deskriptory jsou pouze pro sledování a zatím nejsou podporovány.",
