+++ web-wallet/src/assets/locales/de-de.json
@@ -73,8 +73,7 @@
   "balance_available": "Verfuegbares Guthaben",
   "balance_history": "Guthabenverlauf",
   "base_target": "Basisziel",
-  "bip39_passphrase": "BIP39-Passphrase",
-  "bip39_passphrase_info": "Du kannst optional eine Passphrase hinzufuegen (manchmal als 25. Wort bezeichnet) fuer zusaetzliche Sicherheit. Dies erstellt ein voellig anderes Wallet.",
+  "bip39_passphrase": "BIP39-Passphrase (25. Wort)",
   "bip39_passphrase_warning": "Wenn du eine Passphrase festlegst, MUSST du sie dir merken. Ohne die Passphrase koennen deine Gelder nicht wiederhergestellt werden, selbst mit den 24 Woertern!",
   "bitcoin_pocx_network": "Bitcoin PoCX Netzwerk",
   "block": "Block",
@@ -137,7 +136,7 @@
   "compression": "Kompression",
   "confirm": "Bestaetigen",
   "confirm_backup_written": "Ich habe meine Wiederherstellungsphrase aufgeschrieben und sicher aufbewahrt",
-  "confirm_bip39_passphrase": "Passphrase bestaetigen",
+  "confirm_bip39_passphrase": "25. Wort bestätigen",
   "confirm_send": "Bestaetigen & Senden",
   "confirm_send_message": "Bitte ueberpruefe die Transaktionsdetails vor dem Senden.",
   "confirm_transaction": "Transaktion bestaetigen",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Wallet erstellt. Jeder Mitunterzeichner muss sein Wallet mit denselben Schluesseln erstellen, bevor Geld empfangen wird.",
   "mwallet_address_failed": "Adresse konnte nicht abgerufen werden",
   "mwallet_add_to_contact": "Zu Kontakten hinzufügen",
-  "mwallet_backup_ack": "Ich habe die 24 Wörter aufgeschrieben",
-  "mwallet_backup_title": "Schreibe deine Wiederherstellungsphrase auf",
-  "mwallet_backup_warning": "Diese 24 Wörter sind die einzige Sicherung deines Wallets. Schreibe sie auf Papier und bewahre sie offline auf. Jeder, der sie kennt, kann dein Guthaben ausgeben.",
   "mwallet_branch_legacy": "Legacy (Coin-Type 0')",
   "mwallet_create_failed": "Wallet konnte nicht erstellt werden",
   "mwallet_create_wallet": "Neues Wallet erstellen",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Erstelle ein neues Wallet oder stelle ein vorhandenes aus seiner Wiederherstellungsphrase wieder her.",
   "mwallet_onboarding_title": "Wallet einrichten",
   "mwallet_passphrase_confirm": "Passphrase bestätigen",
-  "mwallet_bip39_toggle": "BIP39-Passphrase hinzufügen (25. Wort)",
-  "mwallet_bip39_label": "BIP39-Passphrase (25. Wort)",
-  "mwallet_bip39_confirm": "25. Wort bestätigen",
-  "mwallet_bip39_warning": "Dieses optionale 25. Wort fließt in Ihre Schlüssel ein. Wenn Sie es setzen, stellt Ihre Wiederherstellungsphrase ALLEIN Ihre Gelder NICHT wieder her — Sie müssen sich auch genau dieses Wort merken. Es ist nicht die Geräte-Passphrase oben.",
-  "mwallet_bip39_restore_warning": "Geben Sie genau das 25. Wort ein, mit dem diese Wallet erstellt wurde. Ein falsches oder fehlendes Wort leitet eine andere, leere Wallet ab.",
+  "bip39_restore_warning": "Geben Sie genau das 25. Wort ein, mit dem diese Wallet erstellt wurde. Ein falsches oder fehlendes Wort leitet eine andere, leere Wallet ab.",
   "mwallet_passphrase_hint": "Verschlüsselt den auf diesem Gerät gespeicherten Seed. Sie ist nicht Teil der Wiederherstellungsphrase — die Wörter allein stellen dein Guthaben immer wieder her.",
   "mwallet_passphrase_label": "Passphrase",
-  "mwallet_passphrase_mismatch": "Passphrasen stimmen nicht überein",
   "mwallet_passphrase_optional": "Passphrase (Optional)",
-  "mwallet_passphrase_title": "Geräte-Passphrase (optional)",
   "mwallet_received": "Empfangen",
   "mwallet_remove_server": "Server entfernen",
   "mwallet_restore_branch": "Ableitungszweig: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Verfügbar",
   "mwallet_title": "Wallet",
   "mwallet_unlock_failed": "Entsperren fehlgeschlagen — falsche Passphrase?",
-  "mwallet_verify_hint": "Gib die abgefragten Wörter aus deiner schriftlichen Sicherung ein.",
-  "mwallet_verify_mismatch": "Die Wörter stimmen nicht mit deiner Wiederherstellungsphrase überein.",
-  "mwallet_verify_title": "Sicherung bestätigen",
   "mwallet_import_bare_key": "Füge einen vollständigen Deskriptor ein (wpkh(...) oder tr(...)), keinen bloßen Schlüssel.",
   "mwallet_import_bare_wif": "Das ist ein bloßer privater WIF-Schlüssel — umschließe ihn als wpkh(DEIN_WIF), um ihn als Einzeladressen-Wallet zu importieren.",
   "mwallet_import_description": "Importiere ein Wallet aus seinen Deskriptoren. Füge einen oder zwei private Deskriptoren ein — öffentliche (xpub) Deskriptoren sind Watch-Only und werden noch nicht unterstützt.",
