+++ web-wallet/src/assets/locales/el.json
@@ -73,8 +73,7 @@
   "balance_available": "Διαθέσιμο υπόλοιπο",
   "balance_history": "Ιστορικό υπολοίπου",
   "base_target": "Βασικός στόχος",
-  "bip39_passphrase": "Κωδικός BIP39",
-  "bip39_passphrase_info": "Μπορείτε να προσθέσετε έναν προαιρετικό κωδικό (μερικές φορές αποκαλείται 25η λέξη) για επιπλέον ασφάλεια. Αυτό δημιουργεί ένα εντελώς διαφορετικό πορτοφόλι.",
+  "bip39_passphrase": "Κωδικός BIP39 (25η λέξη)",
   "bip39_passphrase_warning": "Αν ορίσετε κωδικό, ΠΡΕΠΕΙ να τον θυμάστε. Χωρίς τον κωδικό, τα κεφάλαιά σας δεν μπορούν να ανακτηθούν ακόμα και με τις 24 λέξεις!",
   "bitcoin_pocx_network": "Δίκτυο Bitcoin PoCX",
   "block": "Μπλοκ",
@@ -137,7 +136,7 @@
   "compression": "Συμπίεση",
   "confirm": "Επιβεβαίωση",
   "confirm_backup_written": "Έχω καταγράψει τη φράση ανάκτησης και την έχω αποθηκεύσει με ασφάλεια",
-  "confirm_bip39_passphrase": "Επιβεβαίωση κωδικού",
+  "confirm_bip39_passphrase": "Επιβεβαίωση 25ης λέξης",
   "confirm_send": "Επιβεβαίωση & Αποστολή",
   "confirm_send_message": "Παρακαλώ ελέγξτε τις λεπτομέρειες της συναλλαγής πριν την αποστολή.",
   "confirm_transaction": "Επιβεβαίωση συναλλαγής",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Το πορτοφόλι δημιουργήθηκε. Κάθε συνυπογράφων πρέπει να δημιουργήσει το πορτοφόλι του με τα ίδια κλειδιά πριν από τη λήψη κεφαλαίων.",
   "mwallet_address_failed": "Αποτυχία λήψης διεύθυνσης",
   "mwallet_add_to_contact": "Προσθήκη στις επαφές",
-  "mwallet_backup_ack": "Έχω σημειώσει τις 24 λέξεις",
-  "mwallet_backup_title": "Σημειώστε τη φράση ανάκτησης",
-  "mwallet_backup_warning": "Αυτές οι 24 λέξεις είναι το μοναδικό αντίγραφο ασφαλείας του πορτοφολιού σας. Σημειώστε τις σε χαρτί και φυλάξτε τις εκτός σύνδεσης. Όποιος τις γνωρίζει μπορεί να ξοδέψει τα κεφάλαιά σας.",
   "mwallet_branch_legacy": "παλαιός (coin type 0')",
   "mwallet_create_failed": "Αποτυχία δημιουργίας πορτοφολιού",
   "mwallet_create_wallet": "Δημιουργία νέου πορτοφολιού",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Δημιουργήστε ένα νέο πορτοφόλι ή επαναφέρετε ένα υπάρχον από τη φράση ανάκτησης.",
   "mwallet_onboarding_title": "Ρυθμίστε το πορτοφόλι σας",
   "mwallet_passphrase_confirm": "Επιβεβαίωση κωδικού",
-  "mwallet_bip39_toggle": "Προσθήκη κωδικού BIP39 (25η λέξη)",
-  "mwallet_bip39_label": "Κωδικός BIP39 (25η λέξη)",
-  "mwallet_bip39_confirm": "Επιβεβαίωση 25ης λέξης",
-  "mwallet_bip39_warning": "Αυτή η προαιρετική 25η λέξη ενσωματώνεται στα κλειδιά σας. Αν την ορίσετε, η φράση ανάκτησης ΑΠΟ ΜΟΝΗ της ΔΕΝ θα επαναφέρει τα κεφάλαιά σας — πρέπει να θυμάστε και αυτήν ακριβώς τη λέξη. Δεν είναι ο κωδικός της συσκευής παραπάνω.",
-  "mwallet_bip39_restore_warning": "Εισαγάγετε την ακριβή 25η λέξη με την οποία δημιουργήθηκε αυτό το πορτοφόλι. Λάθος ή ελλιπής λέξη παράγει ένα διαφορετικό, άδειο πορτοφόλι.",
+  "bip39_restore_warning": "Εισαγάγετε την ακριβή 25η λέξη με την οποία δημιουργήθηκε αυτό το πορτοφόλι. Λάθος ή ελλιπής λέξη παράγει ένα διαφορετικό, άδειο πορτοφόλι.",
   "mwallet_passphrase_hint": "Κρυπτογραφεί το seed που αποθηκεύεται σε αυτή τη συσκευή. Δεν αποτελεί μέρος της φράσης ανάκτησης — οι λέξεις από μόνες τους επαναφέρουν πάντα τα κεφάλαιά σας.",
   "mwallet_passphrase_label": "Κωδικός",
-  "mwallet_passphrase_mismatch": "Οι κωδικοί δεν ταιριάζουν",
   "mwallet_passphrase_optional": "Κωδικός (προαιρετικό)",
-  "mwallet_passphrase_title": "Κωδικός συσκευής (προαιρετικό)",
   "mwallet_received": "Ελήφθη",
   "mwallet_remove_server": "Αφαίρεση διακομιστή",
   "mwallet_restore_branch": "Κλάδος παραγωγής: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Διαθέσιμο",
   "mwallet_title": "Πορτοφόλι",
   "mwallet_unlock_failed": "Αδυναμία ξεκλειδώματος — λάθος κωδικός;",
-  "mwallet_verify_hint": "Εισάγετε τις ζητούμενες λέξεις από το γραπτό σας αντίγραφο.",
-  "mwallet_verify_mismatch": "Οι λέξεις δεν ταιριάζουν με τη φράση ανάκτησης.",
-  "mwallet_verify_title": "Επιβεβαιώστε το αντίγραφο ασφαλείας",
   "mwallet_import_bare_key": "Επικολλήστε έναν πλήρη περιγραφέα (wpkh(...) ή tr(...)), όχι σκέτο κλειδί.",
   "mwallet_import_bare_wif": "Αυτό είναι σκέτο ιδιωτικό κλειδί WIF — τυλίξτε το ως wpkh(ΤΟ_WIF_ΣΑΣ) για να το εισαγάγετε ως πορτοφόλι μίας διεύθυνσης.",
   "mwallet_import_description": "Εισαγάγετε ένα πορτοφόλι από τους περιγραφείς του. Επικολλήστε έναν ή δύο ιδιωτικούς περιγραφείς — οι δημόσιοι (xpub) περιγραφείς είναι μόνο για παρακολούθηση και δεν υποστηρίζονται ακόμη.",
