+++ web-wallet/src/assets/locales/en.json
@@ -68,8 +68,7 @@
   "balance_history": "Balance History",
   "base_target": "Base Target",
   "block": "Block",
-  "bip39_passphrase": "BIP39 Passphrase",
-  "bip39_passphrase_info": "You can add an optional passphrase (sometimes called the 25th word) for additional security. This creates a completely different wallet.",
+  "bip39_passphrase": "BIP39 passphrase (25th word)",
   "bip39_passphrase_warning": "If you set a passphrase, you MUST remember it. Without the passphrase, your funds cannot be recovered even with the 24 words!",
   "bitcoin_pocx_network": "Bitcoin PoCX Network",
   "block_details": "Block Details",
@@ -137,7 +136,7 @@
   "compression": "Compression",
   "confirm": "Confirm",
   "confirm_backup_written": "I have written down my recovery phrase and stored it safely",
-  "confirm_bip39_passphrase": "Confirm Passphrase",
+  "confirm_bip39_passphrase": "Confirm 25th word",
   "confirm_send": "Confirm & Send",
   "confirm_send_message": "Please review the transaction details before sending.",
   "confirm_transaction": "Confirm Transaction",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Wallet created. Every co-signer must create their wallet with the same keys before receiving funds.",
   "mwallet_address_failed": "Failed to get an address",
   "mwallet_add_to_contact": "Add to contacts",
-  "mwallet_backup_ack": "I have written down the 24 words",
-  "mwallet_backup_title": "Write down your recovery phrase",
-  "mwallet_backup_warning": "These 24 words are the only backup of your wallet. Write them down on paper and keep them offline. Anyone who knows them can spend your funds.",
   "mwallet_branch_legacy": "legacy (coin type 0')",
   "mwallet_create_failed": "Failed to create the wallet",
   "mwallet_create_wallet": "Create new wallet",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Create a new wallet or restore an existing one from its recovery phrase.",
   "mwallet_onboarding_title": "Set up your wallet",
   "mwallet_passphrase_confirm": "Confirm passphrase",
-  "mwallet_bip39_toggle": "Add BIP39 passphrase (25th word)",
-  "mwallet_bip39_label": "BIP39 passphrase (25th word)",
-  "mwallet_bip39_confirm": "Confirm 25th word",
-  "mwallet_bip39_warning": "This optional 25th word is folded into your keys. If you set it, your recovery phrase ALONE will NOT restore your funds — you must also remember this exact word. It is not the device passphrase above.",
-  "mwallet_bip39_restore_warning": "Enter the exact 25th word this wallet was created with. A wrong or missing word derives a different, empty wallet.",
+  "bip39_restore_warning": "Enter the exact 25th word this wallet was created with. A wrong or missing word derives a different, empty wallet.",
   "mwallet_passphrase_hint": "Encrypts the seed stored on this device. It is not part of the recovery phrase — the words alone always restore your funds.",
   "mwallet_passphrase_label": "Passphrase",
-  "mwallet_passphrase_mismatch": "Passphrases do not match",
   "mwallet_passphrase_optional": "Passphrase (optional)",
-  "mwallet_passphrase_title": "Device passphrase (optional)",
   "mwallet_received": "Received",
   "mwallet_remove_server": "Remove server",
   "mwallet_restore_branch": "Derivation branch: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Spendable",
   "mwallet_title": "Wallet",
   "mwallet_unlock_failed": "Could not unlock — wrong passphrase?",
-  "mwallet_verify_hint": "Enter the requested words from your written backup.",
-  "mwallet_verify_mismatch": "The words do not match your recovery phrase.",
-  "mwallet_verify_title": "Confirm your backup",
   "mwallet_import_bare_key": "Paste a full descriptor (wpkh(...) or tr(...)), not a bare key.",
   "mwallet_import_bare_wif": "That is a bare WIF private key — wrap it as wpkh(YOUR_WIF) to import it as a single-address wallet.",
   "mwallet_import_description": "Import a wallet from its descriptors. Paste one or two private descriptors — public (xpub) descriptors are watch-only and not supported yet.",
