+++ web-wallet/src/assets/locales/es-es.json
@@ -73,8 +73,7 @@
   "balance_available": "Balance disponible",
   "balance_history": "Historial de balance",
   "base_target": "Base Target",
-  "bip39_passphrase": "Frase de contraseña BIP39",
-  "bip39_passphrase_info": "Puede agregar una frase de contraseña opcional (a veces llamada la palabra 25) para mayor seguridad. Esto crea una billetera completamente diferente.",
+  "bip39_passphrase": "Frase de contraseña BIP39 (25.ª palabra)",
   "bip39_passphrase_warning": "Si establece una frase de contraseña, DEBE recordarla. Sin la frase de contraseña, sus fondos no pueden ser recuperados incluso con las 24 palabras.",
   "bitcoin_pocx_network": "Red Bitcoin PoCX",
   "block": "Bloque",
@@ -137,7 +136,7 @@
   "compression": "Compresión",
   "confirm": "Confirmar",
   "confirm_backup_written": "He escrito mi frase de recuperación y la he guardado de forma segura",
-  "confirm_bip39_passphrase": "Confirmar frase de contraseña",
+  "confirm_bip39_passphrase": "Confirmar la 25.ª palabra",
   "confirm_send": "Confirmar y enviar",
   "confirm_send_message": "Por favor revise los detalles de la transacción antes de enviar.",
   "confirm_transaction": "Confirmar transacción",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Billetera creada. Cada cofirmante debe crear su billetera con las mismas claves antes de recibir fondos.",
   "mwallet_address_failed": "No se pudo obtener una dirección",
   "mwallet_add_to_contact": "Añadir a contactos",
-  "mwallet_backup_ack": "He anotado las 24 palabras",
-  "mwallet_backup_title": "Anote su frase de recuperación",
-  "mwallet_backup_warning": "Estas 24 palabras son la única copia de seguridad de su billetera. Anótelas en papel y guárdelas sin conexión. Cualquiera que las conozca puede gastar sus fondos.",
   "mwallet_branch_legacy": "heredada (coin type 0')",
   "mwallet_create_failed": "No se pudo crear la billetera",
   "mwallet_create_wallet": "Crear billetera nueva",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Cree una billetera nueva o restaure una existente a partir de su frase de recuperación.",
   "mwallet_onboarding_title": "Configure su billetera",
   "mwallet_passphrase_confirm": "Confirmar contraseña",
-  "mwallet_bip39_toggle": "Añadir frase de contraseña BIP39 (25.ª palabra)",
-  "mwallet_bip39_label": "Frase de contraseña BIP39 (25.ª palabra)",
-  "mwallet_bip39_confirm": "Confirmar la 25.ª palabra",
-  "mwallet_bip39_warning": "Esta 25.ª palabra opcional se incorpora a tus claves. Si la estableces, la frase de recuperación POR SÍ SOLA NO restaurará tus fondos — también debes recordar esta palabra exacta. No es la contraseña del dispositivo de arriba.",
-  "mwallet_bip39_restore_warning": "Introduce la 25.ª palabra exacta con la que se creó este monedero. Una palabra errónea o ausente deriva un monedero distinto y vacío.",
+  "bip39_restore_warning": "Introduce la 25.ª palabra exacta con la que se creó este monedero. Una palabra errónea o ausente deriva un monedero distinto y vacío.",
   "mwallet_passphrase_hint": "Cifra la semilla almacenada en este dispositivo. No forma parte de la frase de recuperación — las palabras por sí solas siempre restauran sus fondos.",
   "mwallet_passphrase_label": "Contraseña",
-  "mwallet_passphrase_mismatch": "Las contraseñas no coinciden",
   "mwallet_passphrase_optional": "Contraseña (opcional)",
-  "mwallet_passphrase_title": "Contraseña del dispositivo (opcional)",
   "mwallet_received": "Recibido",
   "mwallet_remove_server": "Eliminar servidor",
   "mwallet_restore_branch": "Rama de derivación: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Disponible",
   "mwallet_title": "Billetera",
   "mwallet_unlock_failed": "No se pudo desbloquear — ¿contraseña incorrecta?",
-  "mwallet_verify_hint": "Introduzca las palabras solicitadas desde su copia escrita.",
-  "mwallet_verify_mismatch": "Las palabras no coinciden con su frase de recuperación.",
-  "mwallet_verify_title": "Confirme su copia de seguridad",
   "mwallet_import_bare_key": "Pegue un descriptor completo (wpkh(...) o tr(...)), no una clave suelta.",
   "mwallet_import_bare_wif": "Eso es una clave privada WIF suelta — envuélvala como wpkh(SU_WIF) para importarla como billetera de dirección única.",
   "mwallet_import_description": "Importe una billetera a partir de sus descriptores. Pegue uno o dos descriptores privados — los descriptores públicos (xpub) son de solo observación y aún no están soportados.",
