+++ web-wallet/src/assets/locales/fi.json
@@ -73,8 +73,7 @@
   "balance_available": "Kaytettavissa oleva saldo",
   "balance_history": "Saldohistoria",
   "base_target": "Peruskohde",
-  "bip39_passphrase": "BIP39-salasana",
-  "bip39_passphrase_info": "Voit lisata valinnaisen salasanan (joskus kutsutaan 25. sanaksi) lisaturvallisuuden vuoksi. Tama luo taysin eri lompakon.",
+  "bip39_passphrase": "BIP39-salasana (25. sana)",
   "bip39_passphrase_warning": "Jos asetat salasanan, sinun ON muistettava se. Ilman salasanaa varojasi ei voi palauttaa edes 24 sanalla!",
   "bitcoin_pocx_network": "Bitcoin PoCX -verkko",
   "block": "Lohko",
@@ -137,7 +136,7 @@
   "compression": "Pakkaus",
   "confirm": "Vahvista",
   "confirm_backup_written": "Olen kirjoittanut palautuslauseeni ylos ja sailyttanyt sen turvallisesti",
-  "confirm_bip39_passphrase": "Vahvista salasana",
+  "confirm_bip39_passphrase": "Vahvista 25. sana",
   "confirm_send": "Vahvista ja laheta",
   "confirm_send_message": "Tarkista tapahtuman tiedot ennen lahettamista.",
   "confirm_transaction": "Vahvista tapahtuma",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Lompakko luotu. Jokaisen yhteisallekirjoittajan on luotava lompakkonsa samoilla avaimilla ennen varojen vastaanottamista.",
   "mwallet_address_failed": "Osoitteen haku epäonnistui",
   "mwallet_add_to_contact": "Lisää yhteystietoihin",
-  "mwallet_backup_ack": "Olen kirjoittanut 24 sanaa muistiin",
-  "mwallet_backup_title": "Kirjoita palautuslauseesi muistiin",
-  "mwallet_backup_warning": "Nämä 24 sanaa ovat lompakkosi ainoa varmuuskopio. Kirjoita ne paperille ja säilytä ne verkon ulkopuolella. Kuka tahansa, joka tuntee ne, voi käyttää varasi.",
   "mwallet_branch_legacy": "vanha (coin type 0')",
   "mwallet_create_failed": "Lompakon luonti epäonnistui",
   "mwallet_create_wallet": "Luo uusi lompakko",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Luo uusi lompakko tai palauta olemassa oleva palautuslauseesta.",
   "mwallet_onboarding_title": "Määritä lompakkosi",
   "mwallet_passphrase_confirm": "Vahvista salasana",
-  "mwallet_bip39_toggle": "Lisää BIP39-salasana (25. sana)",
-  "mwallet_bip39_label": "BIP39-salasana (25. sana)",
-  "mwallet_bip39_confirm": "Vahvista 25. sana",
-  "mwallet_bip39_warning": "Tämä valinnainen 25. sana sisällytetään avaimiisi. Jos asetat sen, palautuslause YKSINÄÄN EI palauta varojasi — sinun on muistettava myös tämä tarkka sana. Se ei ole yllä oleva laitteen salasana.",
-  "mwallet_bip39_restore_warning": "Syötä täsmälleen se 25. sana, jolla tämä lompakko luotiin. Väärä tai puuttuva sana johtaa eri, tyhjään lompakkoon.",
+  "bip39_restore_warning": "Syötä täsmälleen se 25. sana, jolla tämä lompakko luotiin. Väärä tai puuttuva sana johtaa eri, tyhjään lompakkoon.",
   "mwallet_passphrase_hint": "Salaa tälle laitteelle tallennetun seedin. Se ei ole osa palautuslausetta — sanat yksinään palauttavat aina varasi.",
   "mwallet_passphrase_label": "Salasana",
-  "mwallet_passphrase_mismatch": "Salasanat eivät täsmää",
   "mwallet_passphrase_optional": "Salasana (valinnainen)",
-  "mwallet_passphrase_title": "Laitteen salasana (valinnainen)",
   "mwallet_received": "Vastaanotettu",
   "mwallet_remove_server": "Poista palvelin",
   "mwallet_restore_branch": "Johdannaishaara: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Käytettävissä",
   "mwallet_title": "Lompakko",
   "mwallet_unlock_failed": "Lukituksen avaus epäonnistui — väärä salasana?",
-  "mwallet_verify_hint": "Syötä pyydetyt sanat kirjoittamastasi varmuuskopiosta.",
-  "mwallet_verify_mismatch": "Sanat eivät vastaa palautuslausettasi.",
-  "mwallet_verify_title": "Vahvista varmuuskopiosi",
   "mwallet_import_bare_key": "Liitä kokonainen deskriptori (wpkh(...) tai tr(...)), ei pelkkää avainta.",
   "mwallet_import_bare_wif": "Tämä on paljas yksityinen WIF-avain — kääri se muotoon wpkh(WIF_AVAIMESI) tuodaksesi sen yhden osoitteen lompakkona.",
   "mwallet_import_description": "Tuo lompakko sen deskriptoreista. Liitä yksi tai kaksi yksityistä deskriptoria — julkiset (xpub) deskriptorit ovat vain seurantaan eikä niitä vielä tueta.",
