+++ web-wallet/src/assets/locales/fr.json
@@ -73,8 +73,7 @@
   "balance_available": "Solde disponible",
   "balance_history": "Historique du solde",
   "base_target": "Cible de base",
-  "bip39_passphrase": "Phrase de passe BIP39",
-  "bip39_passphrase_info": "Vous pouvez ajouter une phrase de passe optionnelle (parfois appelee le 25e mot) pour une securite supplementaire. Cela cree un portefeuille completement different.",
+  "bip39_passphrase": "Phrase de passe BIP39 (25e mot)",
   "bip39_passphrase_warning": "Si vous definissez une phrase de passe, vous DEVEZ vous en souvenir. Sans la phrase de passe, vos fonds ne peuvent pas etre recuperes meme avec les 24 mots !",
   "bitcoin_pocx_network": "Reseau Bitcoin PoCX",
   "block": "Bloc",
@@ -137,7 +136,7 @@
   "compression": "Compression",
   "confirm": "Confirmer",
   "confirm_backup_written": "J'ai ecrit ma phrase de recuperation et je l'ai stockee en securite",
-  "confirm_bip39_passphrase": "Confirmer la phrase de passe",
+  "confirm_bip39_passphrase": "Confirmer le 25e mot",
   "confirm_send": "Confirmer et envoyer",
   "confirm_send_message": "Veuillez verifier les details de la transaction avant d'envoyer.",
   "confirm_transaction": "Confirmer la transaction",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Portefeuille créé. Chaque cosignataire doit créer son portefeuille avec les mêmes clés avant de recevoir des fonds.",
   "mwallet_address_failed": "Impossible d'obtenir une adresse",
   "mwallet_add_to_contact": "Ajouter aux contacts",
-  "mwallet_backup_ack": "J'ai noté les 24 mots",
-  "mwallet_backup_title": "Notez votre phrase de récupération",
-  "mwallet_backup_warning": "Ces 24 mots sont la seule sauvegarde de votre portefeuille. Notez-les sur papier et conservez-les hors ligne. Quiconque les connaît peut dépenser vos fonds.",
   "mwallet_branch_legacy": "héritée (coin type 0')",
   "mwallet_create_failed": "Échec de la création du portefeuille",
   "mwallet_create_wallet": "Créer un nouveau portefeuille",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Créez un nouveau portefeuille ou restaurez-en un existant à partir de sa phrase de récupération.",
   "mwallet_onboarding_title": "Configurer votre portefeuille",
   "mwallet_passphrase_confirm": "Confirmer la phrase secrète",
-  "mwallet_bip39_toggle": "Ajouter une phrase de passe BIP39 (25e mot)",
-  "mwallet_bip39_label": "Phrase de passe BIP39 (25e mot)",
-  "mwallet_bip39_confirm": "Confirmer le 25e mot",
-  "mwallet_bip39_warning": "Ce 25e mot optionnel est intégré à vos clés. Si vous le définissez, votre phrase de récupération SEULE NE restaurera PAS vos fonds — vous devez aussi retenir ce mot exact. Ce n'est pas la phrase de passe de l'appareil ci-dessus.",
-  "mwallet_bip39_restore_warning": "Saisissez le 25e mot exact avec lequel ce portefeuille a été créé. Un mot erroné ou manquant dérive un portefeuille différent et vide.",
+  "bip39_restore_warning": "Saisissez le 25e mot exact avec lequel ce portefeuille a été créé. Un mot erroné ou manquant dérive un portefeuille différent et vide.",
   "mwallet_passphrase_hint": "Chiffre la seed stockée sur cet appareil. Elle ne fait pas partie de la phrase de récupération — les mots seuls restaurent toujours vos fonds.",
   "mwallet_passphrase_label": "Phrase secrète",
-  "mwallet_passphrase_mismatch": "Les phrases secrètes ne correspondent pas",
   "mwallet_passphrase_optional": "Phrase secrète (optionnelle)",
-  "mwallet_passphrase_title": "Phrase secrète de l'appareil (optionnelle)",
   "mwallet_received": "Reçu",
   "mwallet_remove_server": "Supprimer le serveur",
   "mwallet_restore_branch": "Branche de dérivation : {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Disponible",
   "mwallet_title": "Portefeuille",
   "mwallet_unlock_failed": "Déverrouillage impossible — phrase secrète incorrecte ?",
-  "mwallet_verify_hint": "Saisissez les mots demandés depuis votre sauvegarde écrite.",
-  "mwallet_verify_mismatch": "Les mots ne correspondent pas à votre phrase de récupération.",
-  "mwallet_verify_title": "Confirmez votre sauvegarde",
   "mwallet_import_bare_key": "Collez un descripteur complet (wpkh(...) ou tr(...)), pas une clé seule.",
   "mwallet_import_bare_wif": "C'est une clé privée WIF seule — enveloppez-la en wpkh(VOTRE_WIF) pour l'importer comme portefeuille à adresse unique.",
   "mwallet_import_description": "Importez un portefeuille à partir de ses descripteurs. Collez un ou deux descripteurs privés — les descripteurs publics (xpub) sont en lecture seule et ne sont pas encore pris en charge.",
