+++ web-wallet/src/assets/locales/gl.json
@@ -73,8 +73,7 @@
   "balance_available": "Saldo dispoñible",
   "balance_history": "Historial de saldo",
   "base_target": "Obxectivo base",
-  "bip39_passphrase": "Frase de contrasinal BIP39",
-  "bip39_passphrase_info": "Podes engadir unha frase de contrasinal opcional (ás veces chamada a palabra 25) para maior seguridade. Isto crea unha carteira completamente diferente.",
+  "bip39_passphrase": "Frase de contrasinal BIP39 (25.ª palabra)",
   "bip39_passphrase_warning": "Se estableces unha frase de contrasinal, DEBES lembrala. Sen a frase de contrasinal, os teus fondos non se poden recuperar nin coas 24 palabras!",
   "bitcoin_pocx_network": "Rede Bitcoin PoCX",
   "block": "Bloque",
@@ -137,7 +136,7 @@
   "compression": "Compresión",
   "confirm": "Confirmar",
   "confirm_backup_written": "Escribín a miña frase de recuperación e gardeina de forma segura",
-  "confirm_bip39_passphrase": "Confirmar frase de contrasinal",
+  "confirm_bip39_passphrase": "Confirmar a 25.ª palabra",
   "confirm_send": "Confirmar e enviar",
   "confirm_send_message": "Por favor, revisa os detalles da transacción antes de enviar.",
   "confirm_transaction": "Confirmar transacción",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Carteira creada. Cada cofirmante debe crear a súa carteira coas mesmas chaves antes de recibir fondos.",
   "mwallet_address_failed": "Non se puido obter un enderezo",
   "mwallet_add_to_contact": "Engadir aos contactos",
-  "mwallet_backup_ack": "Anotei as 24 palabras",
-  "mwallet_backup_title": "Anota a túa frase de recuperación",
-  "mwallet_backup_warning": "Estas 24 palabras son a única copia de seguridade da túa carteira. Anótaas en papel e gárdaas sen conexión. Calquera que as coñeza pode gastar os teus fondos.",
   "mwallet_branch_legacy": "herdada (coin type 0')",
   "mwallet_create_failed": "Non se puido crear a carteira",
   "mwallet_create_wallet": "Crear carteira nova",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Crea unha carteira nova ou restaura unha existente a partir da frase de recuperación.",
   "mwallet_onboarding_title": "Configura a túa carteira",
   "mwallet_passphrase_confirm": "Confirmar contrasinal",
-  "mwallet_bip39_toggle": "Engadir frase de contrasinal BIP39 (25.ª palabra)",
-  "mwallet_bip39_label": "Frase de contrasinal BIP39 (25.ª palabra)",
-  "mwallet_bip39_confirm": "Confirmar a 25.ª palabra",
-  "mwallet_bip39_warning": "Esta 25.ª palabra opcional incorpórase ás túas claves. Se a estableces, a frase de recuperación POR SI SOA NON restaurará os teus fondos — tamén debes lembrar esta palabra exacta. Non é o contrasinal do dispositivo de arriba.",
-  "mwallet_bip39_restore_warning": "Introduce a 25.ª palabra exacta coa que se creou esta carteira. Unha palabra errónea ou ausente deriva unha carteira distinta e baleira.",
+  "bip39_restore_warning": "Introduce a 25.ª palabra exacta coa que se creou esta carteira. Unha palabra errónea ou ausente deriva unha carteira distinta e baleira.",
   "mwallet_passphrase_hint": "Cifra a semente almacenada neste dispositivo. Non forma parte da frase de recuperación — as palabras por si soas sempre restauran os teus fondos.",
   "mwallet_passphrase_label": "Contrasinal",
-  "mwallet_passphrase_mismatch": "Os contrasinais non coinciden",
   "mwallet_passphrase_optional": "Contrasinal (opcional)",
-  "mwallet_passphrase_title": "Contrasinal do dispositivo (opcional)",
   "mwallet_received": "Recibido",
   "mwallet_remove_server": "Eliminar servidor",
   "mwallet_restore_branch": "Rama de derivación: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Dispoñible",
   "mwallet_title": "Carteira",
   "mwallet_unlock_failed": "Non se puido desbloquear — contrasinal incorrecto?",
-  "mwallet_verify_hint": "Introduce as palabras solicitadas desde a túa copia escrita.",
-  "mwallet_verify_mismatch": "As palabras non coinciden coa túa frase de recuperación.",
-  "mwallet_verify_title": "Confirma a túa copia de seguridade",
   "mwallet_import_bare_key": "Pegue un descritor completo (wpkh(...) ou tr(...)), non unha chave soa.",
   "mwallet_import_bare_wif": "Iso é unha clave privada WIF solta — envólvaa como wpkh(A_SÚA_WIF) para importala como carteira de enderezo único.",
   "mwallet_import_description": "Importe unha carteira a partir dos seus descritores. Pegue un ou dous descritores privados — os descritores públicos (xpub) son só de observación e aínda non están soportados.",
