+++ web-wallet/src/assets/locales/hi.json
@@ -73,8 +73,7 @@
   "balance_available": "उपलब्ध शेष",
   "balance_history": "शेष इतिहास",
   "base_target": "आधार लक्ष्य",
-  "bip39_passphrase": "BIP39 पासफ्रेज",
-  "bip39_passphrase_info": "अतिरिक्त सुरक्षा के लिए आप एक वैकल्पिक पासफ्रेज (कभी-कभी 25वां शब्द कहा जाता है) जोड़ सकते हैं। यह एक पूरी तरह से अलग वॉलेट बनाता है।",
+  "bip39_passphrase": "BIP39 पासफ्रेज (25वाँ शब्द)",
   "bip39_passphrase_warning": "यदि आप पासफ्रेज सेट करते हैं, तो आपको इसे याद रखना होगा। पासफ्रेज के बिना, 24 शब्दों के साथ भी आपके फंड पुनर्प्राप्त नहीं किए जा सकते!",
   "bitcoin_pocx_network": "बिटकॉइन PoCX नेटवर्क",
   "block": "ब्लॉक",
@@ -137,7 +136,7 @@
   "compression": "संपीड़न",
   "confirm": "पुष्टि करें",
   "confirm_backup_written": "मैंने अपना रिकवरी वाक्यांश लिख लिया है और सुरक्षित रूप से संग्रहीत कर लिया है",
-  "confirm_bip39_passphrase": "पासफ्रेज की पुष्टि करें",
+  "confirm_bip39_passphrase": "25वें शब्द की पुष्टि करें",
   "confirm_send": "पुष्टि करें और भेजें",
   "confirm_send_message": "भेजने से पहले लेनदेन विवरण की समीक्षा करें।",
   "confirm_transaction": "लेनदेन की पुष्टि करें",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "वॉलेट बना दिया गया। फंड प्राप्त करने से पहले प्रत्येक सह-हस्ताक्षरकर्ता को समान कुंजियों के साथ अपना वॉलेट बनाना होगा।",
   "mwallet_address_failed": "पता प्राप्त करने में विफल",
   "mwallet_add_to_contact": "संपर्कों में जोड़ें",
-  "mwallet_backup_ack": "मैंने 24 शब्द लिख लिए हैं",
-  "mwallet_backup_title": "अपना रिकवरी वाक्यांश लिख लें",
-  "mwallet_backup_warning": "ये 24 शब्द आपके वॉलेट का एकमात्र बैकअप हैं। इन्हें कागज पर लिखें और ऑफलाइन रखें। जो कोई भी इन्हें जानता है, वह आपकी धनराशि खर्च कर सकता है।",
   "mwallet_branch_legacy": "पुरानी (coin type 0')",
   "mwallet_create_failed": "वॉलेट बनाने में विफल",
   "mwallet_create_wallet": "नया वॉलेट बनाएं",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "नया वॉलेट बनाएं या रिकवरी वाक्यांश से मौजूदा वॉलेट पुनर्स्थापित करें।",
   "mwallet_onboarding_title": "अपना वॉलेट सेट करें",
   "mwallet_passphrase_confirm": "पासफ़्रेज़ की पुष्टि करें",
-  "mwallet_bip39_toggle": "BIP39 पासफ्रेज जोड़ें (25वाँ शब्द)",
-  "mwallet_bip39_label": "BIP39 पासफ्रेज (25वाँ शब्द)",
-  "mwallet_bip39_confirm": "25वें शब्द की पुष्टि करें",
-  "mwallet_bip39_warning": "यह वैकल्पिक 25वाँ शब्द आपकी कुंजियों में शामिल हो जाता है। यदि आप इसे सेट करते हैं, तो केवल रिकवरी फ्रेज से आपकी धनराशि बहाल नहीं होगी — आपको यह सटीक शब्द भी याद रखना होगा। यह ऊपर वाला डिवाइस पासफ्रेज नहीं है।",
-  "mwallet_bip39_restore_warning": "वही 25वाँ शब्द दर्ज करें जिसके साथ यह वॉलेट बनाया गया था। गलत या अनुपस्थित शब्द एक अलग, खाली वॉलेट व्युत्पन्न करता है।",
+  "bip39_restore_warning": "वही 25वाँ शब्द दर्ज करें जिसके साथ यह वॉलेट बनाया गया था। गलत या अनुपस्थित शब्द एक अलग, खाली वॉलेट व्युत्पन्न करता है।",
   "mwallet_passphrase_hint": "इस डिवाइस पर संग्रहीत सीड को एन्क्रिप्ट करता है। यह रिकवरी वाक्यांश का हिस्सा नहीं है — केवल शब्द ही हमेशा आपकी धनराशि पुनर्स्थापित करते हैं।",
   "mwallet_passphrase_label": "पासफ़्रेज़",
-  "mwallet_passphrase_mismatch": "पासफ़्रेज़ मेल नहीं खाते",
   "mwallet_passphrase_optional": "पासफ़्रेज़ (वैकल्पिक)",
-  "mwallet_passphrase_title": "डिवाइस पासफ़्रेज़ (वैकल्पिक)",
   "mwallet_received": "प्राप्त हुआ",
   "mwallet_remove_server": "सर्वर हटाएं",
   "mwallet_restore_branch": "डेरिवेशन शाखा: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "खर्च योग्य",
   "mwallet_title": "वॉलेट",
   "mwallet_unlock_failed": "अनलॉक नहीं हो सका — गलत पासफ़्रेज़?",
-  "mwallet_verify_hint": "अपने लिखित बैकअप से मांगे गए शब्द दर्ज करें।",
-  "mwallet_verify_mismatch": "शब्द आपके रिकवरी वाक्यांश से मेल नहीं खाते।",
-  "mwallet_verify_title": "अपने बैकअप की पुष्टि करें",
   "mwallet_import_bare_key": "पूरा डिस्क्रिप्टर चिपकाएँ (wpkh(...) या tr(...)), केवल कुंजी नहीं।",
   "mwallet_import_bare_wif": "यह एक अकेली WIF निजी कुंजी है — इसे wpkh(आपकी_WIF) के रूप में लपेटें ताकि इसे एकल-पता वॉलेट के रूप में आयात किया जा सके।",
   "mwallet_import_description": "वॉलेट को उसके डिस्क्रिप्टरों से आयात करें। एक या दो निजी डिस्क्रिप्टर चिपकाएँ — सार्वजनिक (xpub) डिस्क्रिप्टर केवल-देखने वाले होते हैं और अभी समर्थित नहीं हैं।",
