+++ web-wallet/src/assets/locales/hr.json
@@ -73,8 +73,7 @@
   "balance_available": "Dostupno stanje",
   "balance_history": "Povijest stanja",
   "base_target": "Osnovni cilj",
-  "bip39_passphrase": "BIP39 lozinka",
-  "bip39_passphrase_info": "Mozete dodati neobaveznu lozinku (ponekad zvanu 25. rijec) za dodatnu sigurnost. Ovo stvara potpuno drugaciji novcanik.",
+  "bip39_passphrase": "BIP39 lozinka (25. riječ)",
   "bip39_passphrase_warning": "Ako postavite lozinku, MORATE je zapamtiti. Bez lozinke, sredstva se ne mogu oporaviti cak ni s 24 rijeci!",
   "bitcoin_pocx_network": "Bitcoin PoCX mreza",
   "block": "Blok",
@@ -137,7 +136,7 @@
   "compression": "Kompresija",
   "confirm": "Potvrdi",
   "confirm_backup_written": "Zapisao/la sam frazu za oporavak i sigurno je pohranio/la",
-  "confirm_bip39_passphrase": "Potvrdi lozinku",
+  "confirm_bip39_passphrase": "Potvrdi 25. riječ",
   "confirm_send": "Potvrdi i posalji",
   "confirm_send_message": "Molimo provjerite detalje transakcije prije slanja.",
   "confirm_transaction": "Potvrdi transakciju",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Novčanik kreiran. Svaki supotpisnik mora kreirati svoj novčanik s istim ključevima prije primanja sredstava.",
   "mwallet_address_failed": "Nije moguće dohvatiti adresu",
   "mwallet_add_to_contact": "Dodaj u kontakte",
-  "mwallet_backup_ack": "Zapisao sam 24 riječi",
-  "mwallet_backup_title": "Zapišite svoju frazu za oporavak",
-  "mwallet_backup_warning": "Ove 24 riječi jedina su sigurnosna kopija vašeg novčanika. Zapišite ih na papir i čuvajte izvan mreže. Svatko tko ih zna može potrošiti vaša sredstva.",
   "mwallet_branch_legacy": "naslijeđena (coin type 0')",
   "mwallet_create_failed": "Novčanik nije moguće kreirati",
   "mwallet_create_wallet": "Kreiraj novi novčanik",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Kreirajte novi novčanik ili vratite postojeći iz fraze za oporavak.",
   "mwallet_onboarding_title": "Postavite svoj novčanik",
   "mwallet_passphrase_confirm": "Potvrdi lozinku",
-  "mwallet_bip39_toggle": "Dodaj BIP39 lozinku (25. riječ)",
-  "mwallet_bip39_label": "BIP39 lozinka (25. riječ)",
-  "mwallet_bip39_confirm": "Potvrdi 25. riječ",
-  "mwallet_bip39_warning": "Ova neobavezna 25. riječ ugrađuje se u vaše ključeve. Ako je postavite, fraza za oporavak SAMA NEĆE vratiti vaša sredstva — morate zapamtiti i točno ovu riječ. To nije lozinka uređaja iznad.",
-  "mwallet_bip39_restore_warning": "Unesite točnu 25. riječ s kojom je ovaj novčanik stvoren. Pogrešna ili izostavljena riječ izvodi drugi, prazan novčanik.",
+  "bip39_restore_warning": "Unesite točnu 25. riječ s kojom je ovaj novčanik stvoren. Pogrešna ili izostavljena riječ izvodi drugi, prazan novčanik.",
   "mwallet_passphrase_hint": "Šifrira seed pohranjen na ovom uređaju. Nije dio fraze za oporavak — same riječi uvijek vraćaju vaša sredstva.",
   "mwallet_passphrase_label": "Lozinka",
-  "mwallet_passphrase_mismatch": "Lozinke se ne podudaraju",
   "mwallet_passphrase_optional": "Lozinka (neobavezno)",
-  "mwallet_passphrase_title": "Lozinka uređaja (neobavezno)",
   "mwallet_received": "Primljeno",
   "mwallet_remove_server": "Ukloni poslužitelj",
   "mwallet_restore_branch": "Grana derivacije: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Raspoloživo",
   "mwallet_title": "Novčanik",
   "mwallet_unlock_failed": "Otključavanje nije uspjelo — pogrešna lozinka?",
-  "mwallet_verify_hint": "Unesite tražene riječi iz zapisane kopije.",
-  "mwallet_verify_mismatch": "Riječi se ne podudaraju s vašom frazom za oporavak.",
-  "mwallet_verify_title": "Potvrdite sigurnosnu kopiju",
   "mwallet_import_bare_key": "Zalijepite potpuni deskriptor (wpkh(...) ili tr(...)), a ne goli ključ.",
   "mwallet_import_bare_wif": "To je goli privatni WIF ključ — omotajte ga kao wpkh(VAŠ_WIF) da biste ga uvezli kao novčanik s jednom adresom.",
   "mwallet_import_description": "Uvezite novčanik iz njegovih deskriptora. Zalijepite jedan ili dva privatna deskriptora — javni (xpub) deskriptori su samo za promatranje i još nisu podržani.",
