+++ web-wallet/src/assets/locales/id.json
@@ -73,8 +73,7 @@
   "balance_available": "Saldo Tersedia",
   "balance_history": "Riwayat Saldo",
   "base_target": "Target Dasar",
-  "bip39_passphrase": "Frasa Sandi BIP39",
-  "bip39_passphrase_info": "Anda dapat menambahkan frasa sandi opsional (kadang disebut kata ke-25) untuk keamanan tambahan. Ini membuat dompet yang sama sekali berbeda.",
+  "bip39_passphrase": "Frasa Sandi BIP39 (kata ke-25)",
   "bip39_passphrase_warning": "Jika Anda menetapkan frasa sandi, Anda HARUS mengingatnya. Tanpa frasa sandi, dana Anda tidak dapat dipulihkan bahkan dengan 24 kata!",
   "bitcoin_pocx_network": "Jaringan Bitcoin PoCX",
   "block": "Blok",
@@ -137,7 +136,7 @@
   "compression": "Kompresi",
   "confirm": "Konfirmasi",
   "confirm_backup_written": "Saya telah menulis frasa pemulihan saya dan menyimpannya dengan aman",
-  "confirm_bip39_passphrase": "Konfirmasi Frasa Sandi",
+  "confirm_bip39_passphrase": "Konfirmasi kata ke-25",
   "confirm_send": "Konfirmasi & Kirim",
   "confirm_send_message": "Tinjau detail transaksi sebelum mengirim.",
   "confirm_transaction": "Konfirmasi Transaksi",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Dompet dibuat. Setiap rekan penandatangan harus membuat dompet mereka dengan kunci yang sama sebelum menerima dana.",
   "mwallet_address_failed": "Gagal mendapatkan alamat",
   "mwallet_add_to_contact": "Tambahkan ke kontak",
-  "mwallet_backup_ack": "Saya telah menulis 24 kata tersebut",
-  "mwallet_backup_title": "Tulis frasa pemulihan Anda",
-  "mwallet_backup_warning": "24 kata ini adalah satu-satunya cadangan dompet Anda. Tulis di kertas dan simpan secara offline. Siapa pun yang mengetahuinya dapat membelanjakan dana Anda.",
   "mwallet_branch_legacy": "lama (coin type 0')",
   "mwallet_create_failed": "Gagal membuat dompet",
   "mwallet_create_wallet": "Buat dompet baru",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Buat dompet baru atau pulihkan yang sudah ada dari frasa pemulihannya.",
   "mwallet_onboarding_title": "Siapkan dompet Anda",
   "mwallet_passphrase_confirm": "Konfirmasi frasa sandi",
-  "mwallet_bip39_toggle": "Tambahkan Frasa Sandi BIP39 (kata ke-25)",
-  "mwallet_bip39_label": "Frasa Sandi BIP39 (kata ke-25)",
-  "mwallet_bip39_confirm": "Konfirmasi kata ke-25",
-  "mwallet_bip39_warning": "Kata ke-25 opsional ini menyatu ke dalam kunci Anda. Jika Anda mengaturnya, frasa pemulihan SAJA TIDAK akan memulihkan dana Anda — Anda juga harus mengingat kata persis ini. Ini bukan frasa sandi perangkat di atas.",
-  "mwallet_bip39_restore_warning": "Masukkan kata ke-25 persis seperti saat dompet ini dibuat. Kata yang salah atau hilang menurunkan dompet lain yang kosong.",
+  "bip39_restore_warning": "Masukkan kata ke-25 persis seperti saat dompet ini dibuat. Kata yang salah atau hilang menurunkan dompet lain yang kosong.",
   "mwallet_passphrase_hint": "Mengenkripsi seed yang tersimpan di perangkat ini. Bukan bagian dari frasa pemulihan — kata-kata itu sendiri selalu memulihkan dana Anda.",
   "mwallet_passphrase_label": "Frasa sandi",
-  "mwallet_passphrase_mismatch": "Frasa sandi tidak cocok",
   "mwallet_passphrase_optional": "Frasa sandi (opsional)",
-  "mwallet_passphrase_title": "Frasa sandi perangkat (opsional)",
   "mwallet_received": "Diterima",
   "mwallet_remove_server": "Hapus server",
   "mwallet_restore_branch": "Cabang derivasi: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Dapat dibelanjakan",
   "mwallet_title": "Dompet",
   "mwallet_unlock_failed": "Tidak dapat membuka kunci — frasa sandi salah?",
-  "mwallet_verify_hint": "Masukkan kata-kata yang diminta dari cadangan tertulis Anda.",
-  "mwallet_verify_mismatch": "Kata-kata tidak cocok dengan frasa pemulihan Anda.",
-  "mwallet_verify_title": "Konfirmasi cadangan Anda",
   "mwallet_import_bare_key": "Tempel deskriptor lengkap (wpkh(...) atau tr(...)), bukan kunci saja.",
   "mwallet_import_bare_wif": "Itu adalah kunci privat WIF polos — bungkus sebagai wpkh(WIF_ANDA) untuk mengimpornya sebagai dompet satu alamat.",
   "mwallet_import_description": "Impor dompet dari deskriptornya. Tempel satu atau dua deskriptor privat — deskriptor publik (xpub) bersifat hanya-pantau dan belum didukung.",
