+++ web-wallet/src/assets/locales/it.json
@@ -73,8 +73,7 @@
   "balance_available": "Saldo disponibile",
   "balance_history": "Storico saldo",
   "base_target": "Obiettivo base",
-  "bip39_passphrase": "Passphrase BIP39",
-  "bip39_passphrase_info": "Puoi aggiungere una passphrase opzionale (a volte chiamata 25a parola) per maggiore sicurezza. Questo crea un portafoglio completamente diverso.",
+  "bip39_passphrase": "Passphrase BIP39 (25ª parola)",
   "bip39_passphrase_warning": "Se imposti una passphrase, DEVI ricordarla. Senza la passphrase, i tuoi fondi non possono essere recuperati nemmeno con le 24 parole!",
   "bitcoin_pocx_network": "Rete Bitcoin PoCX",
   "block": "Blocco",
@@ -137,7 +136,7 @@
   "compression": "Compressione",
   "confirm": "Conferma",
   "confirm_backup_written": "Ho scritto la mia frase di recupero e l'ho conservata in modo sicuro",
-  "confirm_bip39_passphrase": "Conferma passphrase",
+  "confirm_bip39_passphrase": "Conferma la 25ª parola",
   "confirm_send": "Conferma e invia",
   "confirm_send_message": "Verifica i dettagli della transazione prima di inviare.",
   "confirm_transaction": "Conferma transazione",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Wallet creato. Ogni cofirmatario deve creare il proprio wallet con le stesse chiavi prima di ricevere fondi.",
   "mwallet_address_failed": "Impossibile ottenere un indirizzo",
   "mwallet_add_to_contact": "Aggiungi ai contatti",
-  "mwallet_backup_ack": "Ho annotato le 24 parole",
-  "mwallet_backup_title": "Annota la tua frase di recupero",
-  "mwallet_backup_warning": "Queste 24 parole sono l'unico backup del tuo portafoglio. Annotale su carta e conservale offline. Chiunque le conosca può spendere i tuoi fondi.",
   "mwallet_branch_legacy": "legacy (coin type 0')",
   "mwallet_create_failed": "Impossibile creare il portafoglio",
   "mwallet_create_wallet": "Crea nuovo portafoglio",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Crea un nuovo portafoglio o ripristinane uno esistente dalla sua frase di recupero.",
   "mwallet_onboarding_title": "Configura il tuo portafoglio",
   "mwallet_passphrase_confirm": "Conferma passphrase",
-  "mwallet_bip39_toggle": "Aggiungi passphrase BIP39 (25ª parola)",
-  "mwallet_bip39_label": "Passphrase BIP39 (25ª parola)",
-  "mwallet_bip39_confirm": "Conferma la 25ª parola",
-  "mwallet_bip39_warning": "Questa 25ª parola opzionale viene incorporata nelle tue chiavi. Se la imposti, la frase di recupero DA SOLA NON ripristinerà i tuoi fondi — devi ricordare anche questa parola esatta. Non è la passphrase del dispositivo qui sopra.",
-  "mwallet_bip39_restore_warning": "Inserisci la 25ª parola esatta con cui è stato creato questo portafoglio. Una parola errata o mancante deriva un portafoglio diverso e vuoto.",
+  "bip39_restore_warning": "Inserisci la 25ª parola esatta con cui è stato creato questo portafoglio. Una parola errata o mancante deriva un portafoglio diverso e vuoto.",
   "mwallet_passphrase_hint": "Cifra il seed memorizzato su questo dispositivo. Non fa parte della frase di recupero — le parole da sole ripristinano sempre i tuoi fondi.",
   "mwallet_passphrase_label": "Passphrase",
-  "mwallet_passphrase_mismatch": "Le passphrase non corrispondono",
   "mwallet_passphrase_optional": "Passphrase (opzionale)",
-  "mwallet_passphrase_title": "Passphrase del dispositivo (opzionale)",
   "mwallet_received": "Ricevuto",
   "mwallet_remove_server": "Rimuovi server",
   "mwallet_restore_branch": "Ramo di derivazione: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Disponibile",
   "mwallet_title": "Portafoglio",
   "mwallet_unlock_failed": "Sblocco non riuscito — passphrase errata?",
-  "mwallet_verify_hint": "Inserisci le parole richieste dal tuo backup scritto.",
-  "mwallet_verify_mismatch": "Le parole non corrispondono alla tua frase di recupero.",
-  "mwallet_verify_title": "Conferma il tuo backup",
   "mwallet_import_bare_key": "Incolla un descrittore completo (wpkh(...) o tr(...)), non una chiave semplice.",
   "mwallet_import_bare_wif": "Questa è una chiave privata WIF nuda — avvolgila come wpkh(TUA_WIF) per importarla come portafoglio a indirizzo singolo.",
   "mwallet_import_description": "Importa un portafoglio dai suoi descrittori. Incolla uno o due descrittori privati — i descrittori pubblici (xpub) sono di sola osservazione e non sono ancora supportati.",
