+++ web-wallet/src/assets/locales/ja.json
@@ -73,8 +73,7 @@
   "balance_available": "利用可能残高",
   "balance_history": "残高履歴",
   "base_target": "ベースターゲット",
-  "bip39_passphrase": "BIP39パスフレーズ",
-  "bip39_passphrase_info": "追加のセキュリティのために、オプションのパスフレーズ（25番目の単語とも呼ばれます）を追加できます。これにより、完全に異なるウォレットが作成されます。",
+  "bip39_passphrase": "BIP39パスフレーズ（25番目の単語）",
   "bip39_passphrase_warning": "パスフレーズを設定した場合、必ず覚えておく必要があります。パスフレーズがないと、24単語があっても資金を回復できません！",
   "bitcoin_pocx_network": "Bitcoin PoCXネットワーク",
   "block": "ブロック",
@@ -137,7 +136,7 @@
   "compression": "圧縮",
   "confirm": "確認",
   "confirm_backup_written": "リカバリーフレーズを書き留めて安全に保管しました",
-  "confirm_bip39_passphrase": "パスフレーズを確認",
+  "confirm_bip39_passphrase": "25番目の単語を確認",
   "confirm_send": "確認して送信",
   "confirm_send_message": "送信前にトランザクションの詳細を確認してください。",
   "confirm_transaction": "トランザクションを確認",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "ウォレットを作成しました。資金を受け取る前に、すべての共同署名者が同じ鍵でウォレットを作成する必要があります。",
   "mwallet_address_failed": "アドレスの取得に失敗しました",
   "mwallet_add_to_contact": "連絡先に追加",
-  "mwallet_backup_ack": "24語を書き留めました",
-  "mwallet_backup_title": "リカバリーフレーズを書き留めてください",
-  "mwallet_backup_warning": "この24語はウォレットの唯一のバックアップです。紙に書き留めてオフラインで保管してください。これらを知っている人は誰でもあなたの資金を使うことができます。",
   "mwallet_branch_legacy": "レガシー（coin type 0'）",
   "mwallet_create_failed": "ウォレットの作成に失敗しました",
   "mwallet_create_wallet": "新しいウォレットを作成",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "新しいウォレットを作成するか、リカバリーフレーズから既存のウォレットを復元します。",
   "mwallet_onboarding_title": "ウォレットをセットアップ",
   "mwallet_passphrase_confirm": "パスフレーズの確認",
-  "mwallet_bip39_toggle": "BIP39パスフレーズを追加（25番目の単語）",
-  "mwallet_bip39_label": "BIP39パスフレーズ（25番目の単語）",
-  "mwallet_bip39_confirm": "25番目の単語を確認",
-  "mwallet_bip39_warning": "この任意の25番目の単語は鍵に組み込まれます。設定した場合、リカバリーフレーズだけでは資金を復元できません — この単語も正確に覚えておく必要があります。上のデバイスパスフレーズとは別のものです。",
-  "mwallet_bip39_restore_warning": "このウォレット作成時に使った25番目の単語を正確に入力してください。単語が違う、または欠けていると、別の空のウォレットが導出されます。",
+  "bip39_restore_warning": "このウォレット作成時に使った25番目の単語を正確に入力してください。単語が違う、または欠けていると、別の空のウォレットが導出されます。",
   "mwallet_passphrase_hint": "このデバイスに保存されるシードを暗号化します。リカバリーフレーズの一部ではありません — 単語だけで常に資金を復元できます。",
   "mwallet_passphrase_label": "パスフレーズ",
-  "mwallet_passphrase_mismatch": "パスフレーズが一致しません",
   "mwallet_passphrase_optional": "パスフレーズ（任意）",
-  "mwallet_passphrase_title": "デバイスパスフレーズ（任意）",
   "mwallet_received": "受取",
   "mwallet_remove_server": "サーバーを削除",
   "mwallet_restore_branch": "導出ブランチ: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "利用可能",
   "mwallet_title": "ウォレット",
   "mwallet_unlock_failed": "ロック解除できませんでした — パスフレーズが違いますか？",
-  "mwallet_verify_hint": "書き留めたバックアップから指定された単語を入力してください。",
-  "mwallet_verify_mismatch": "単語がリカバリーフレーズと一致しません。",
-  "mwallet_verify_title": "バックアップの確認",
   "mwallet_import_bare_key": "鍵単体ではなく、完全なディスクリプタ（wpkh(...) または tr(...)）を貼り付けてください。",
   "mwallet_import_bare_wif": "これは生の WIF 秘密鍵です — wpkh(あなたのWIF) の形で囲むと、単一アドレスのウォレットとしてインポートできます。",
   "mwallet_import_description": "ディスクリプタからウォレットをインポートします。秘密ディスクリプタを 1 つまたは 2 つ貼り付けてください — 公開（xpub）ディスクリプタはウォッチオンリーのため、まだサポートされていません。",
