+++ web-wallet/src/assets/locales/lt.json
@@ -73,8 +73,7 @@
   "balance_available": "Turimas balansas",
   "balance_history": "Balanso istorija",
   "base_target": "Bazinis taikinys",
-  "bip39_passphrase": "BIP39 slaptazodis",
-  "bip39_passphrase_info": "Galite prideti neprivaloma slaptazodi (kartais vadinama 25-uoju zodziu) papildomam saugumui. Tai sukuria visai kita pinigine.",
+  "bip39_passphrase": "BIP39 slaptažodis (25-as žodis)",
   "bip39_passphrase_warning": "Jei nustatysite slaptazodi, TURITE ji atsiminti. Be slaptazodzio lesu negalima atkurti net su 24 zodziais!",
   "bitcoin_pocx_network": "Bitcoin PoCX tinklas",
   "block": "Blokas",
@@ -137,7 +136,7 @@
   "compression": "Suspaudimas",
   "confirm": "Patvirtinti",
   "confirm_backup_written": "Uzrasiau atkurimo fraze ir saugiai ja issaugojau",
-  "confirm_bip39_passphrase": "Patvirtinti slaptazodi",
+  "confirm_bip39_passphrase": "Patvirtinti 25-ą žodį",
   "confirm_send": "Patvirtinti ir siusti",
   "confirm_send_message": "Prasome patikrinti operacijos detales pries siuntima.",
   "confirm_transaction": "Patvirtinti operacija",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Piniginė sukurta. Kiekvienas bendrapasirašytojas turi sukurti savo piniginę su tais pačiais raktais prieš gaudamas lėšas.",
   "mwallet_address_failed": "Nepavyko gauti adreso",
   "mwallet_add_to_contact": "Pridėti prie kontaktų",
-  "mwallet_backup_ack": "Užsirašiau 24 žodžius",
-  "mwallet_backup_title": "Užsirašykite savo atkūrimo frazę",
-  "mwallet_backup_warning": "Šie 24 žodžiai yra vienintelė jūsų piniginės atsarginė kopija. Užsirašykite juos ant popieriaus ir laikykite neprisijungę. Bet kas, kas juos žino, gali išleisti jūsų lėšas.",
   "mwallet_branch_legacy": "senoji (coin type 0')",
   "mwallet_create_failed": "Nepavyko sukurti piniginės",
   "mwallet_create_wallet": "Sukurti naują piniginę",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Sukurkite naują piniginę arba atkurkite esamą iš atkūrimo frazės.",
   "mwallet_onboarding_title": "Nustatykite savo piniginę",
   "mwallet_passphrase_confirm": "Patvirtinti slaptafrazę",
-  "mwallet_bip39_toggle": "Pridėti BIP39 slaptažodį (25-ą žodį)",
-  "mwallet_bip39_label": "BIP39 slaptažodis (25-as žodis)",
-  "mwallet_bip39_confirm": "Patvirtinti 25-ą žodį",
-  "mwallet_bip39_warning": "Šis neprivalomas 25-as žodis įtraukiamas į jūsų raktus. Jį nustačius, atkūrimo frazė VIENA jūsų lėšų NEATKURS — turite atsiminti ir būtent šį žodį. Tai nėra aukščiau esantis įrenginio slaptažodis.",
-  "mwallet_bip39_restore_warning": "Įveskite tikslų 25-ą žodį, su kuriuo ši piniginė buvo sukurta. Neteisingas ar praleistas žodis išveda kitą, tuščią piniginę.",
+  "bip39_restore_warning": "Įveskite tikslų 25-ą žodį, su kuriuo ši piniginė buvo sukurta. Neteisingas ar praleistas žodis išveda kitą, tuščią piniginę.",
   "mwallet_passphrase_hint": "Užšifruoja šiame įrenginyje saugomą seed. Ji nėra atkūrimo frazės dalis — vien žodžiai visada atkuria jūsų lėšas.",
   "mwallet_passphrase_label": "Slaptafrazė",
-  "mwallet_passphrase_mismatch": "Slaptafrazės nesutampa",
   "mwallet_passphrase_optional": "Slaptafrazė (nebūtina)",
-  "mwallet_passphrase_title": "Įrenginio slaptafrazė (nebūtina)",
   "mwallet_received": "Gauta",
   "mwallet_remove_server": "Pašalinti serverį",
   "mwallet_restore_branch": "Išvedimo šaka: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Išleidžiama",
   "mwallet_title": "Piniginė",
   "mwallet_unlock_failed": "Nepavyko atrakinti — neteisinga slaptafrazė?",
-  "mwallet_verify_hint": "Įveskite prašomus žodžius iš savo užrašytos kopijos.",
-  "mwallet_verify_mismatch": "Žodžiai nesutampa su jūsų atkūrimo fraze.",
-  "mwallet_verify_title": "Patvirtinkite atsarginę kopiją",
   "mwallet_import_bare_key": "Įklijuokite pilną deskriptorių (wpkh(...) arba tr(...)), o ne pliką raktą.",
   "mwallet_import_bare_wif": "Tai plikas privatus WIF raktas — apgaubkite jį kaip wpkh(JŪSŲ_WIF), kad importuotumėte kaip vieno adreso piniginę.",
   "mwallet_import_description": "Importuokite piniginę iš jos deskriptorių. Įklijuokite vieną arba du privačius deskriptorius — vieši (xpub) deskriptoriai yra tik stebėjimui ir dar nepalaikomi.",
