+++ web-wallet/src/assets/locales/nl.json
@@ -73,8 +73,7 @@
   "balance_available": "Beschikbaar saldo",
   "balance_history": "Saldo geschiedenis",
   "base_target": "Basisdoel",
-  "bip39_passphrase": "BIP39 Wachtwoordzin",
-  "bip39_passphrase_info": "U kunt een optionele wachtwoordzin toevoegen (soms het 25e woord genoemd) voor extra beveiliging. Dit creëert een volledig andere portemonnee.",
+  "bip39_passphrase": "BIP39-wachtwoordzin (25e woord)",
   "bip39_passphrase_warning": "Als u een wachtwoordzin instelt, MOET u deze onthouden. Zonder de wachtwoordzin kunnen uw fondsen niet worden hersteld, zelfs niet met de 24 woorden!",
   "bitcoin_pocx_network": "Bitcoin PoCX Netwerk",
   "block": "Blok",
@@ -137,7 +136,7 @@
   "compression": "Compressie",
   "confirm": "Bevestigen",
   "confirm_backup_written": "Ik heb mijn herstelzin opgeschreven en veilig bewaard",
-  "confirm_bip39_passphrase": "Bevestig Wachtwoordzin",
+  "confirm_bip39_passphrase": "Bevestig het 25e woord",
   "confirm_send": "Bevestigen & Versturen",
   "confirm_send_message": "Controleer de transactiedetails voordat u verstuurt.",
   "confirm_transaction": "Transactie Bevestigen",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Portemonnee aangemaakt. Elke mede-ondertekenaar moet zijn portemonnee met dezelfde sleutels aanmaken voordat er geld wordt ontvangen.",
   "mwallet_address_failed": "Kon geen adres ophalen",
   "mwallet_add_to_contact": "Toevoegen aan contacten",
-  "mwallet_backup_ack": "Ik heb de 24 woorden opgeschreven",
-  "mwallet_backup_title": "Schrijf uw herstelzin op",
-  "mwallet_backup_warning": "Deze 24 woorden zijn de enige back-up van uw portemonnee. Schrijf ze op papier en bewaar ze offline. Iedereen die ze kent kan uw tegoed uitgeven.",
   "mwallet_branch_legacy": "legacy (coin type 0')",
   "mwallet_create_failed": "Kon de portemonnee niet aanmaken",
   "mwallet_create_wallet": "Nieuwe portemonnee aanmaken",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Maak een nieuwe portemonnee aan of herstel een bestaande vanuit de herstelzin.",
   "mwallet_onboarding_title": "Stel uw portemonnee in",
   "mwallet_passphrase_confirm": "Wachtwoordzin bevestigen",
-  "mwallet_bip39_toggle": "BIP39-wachtwoordzin toevoegen (25e woord)",
-  "mwallet_bip39_label": "BIP39-wachtwoordzin (25e woord)",
-  "mwallet_bip39_confirm": "Bevestig het 25e woord",
-  "mwallet_bip39_warning": "Dit optionele 25e woord wordt in je sleutels verwerkt. Als je het instelt, herstelt je herstelzin ALLEEN je tegoeden NIET — je moet ook precies dit woord onthouden. Het is niet de apparaatwachtwoordzin hierboven.",
-  "mwallet_bip39_restore_warning": "Voer exact het 25e woord in waarmee deze wallet is aangemaakt. Een verkeerd of ontbrekend woord leidt tot een andere, lege wallet.",
+  "bip39_restore_warning": "Voer exact het 25e woord in waarmee deze wallet is aangemaakt. Een verkeerd of ontbrekend woord leidt tot een andere, lege wallet.",
   "mwallet_passphrase_hint": "Versleutelt de seed die op dit apparaat is opgeslagen. Het maakt geen deel uit van de herstelzin — de woorden alleen herstellen altijd uw tegoed.",
   "mwallet_passphrase_label": "Wachtwoordzin",
-  "mwallet_passphrase_mismatch": "Wachtwoordzinnen komen niet overeen",
   "mwallet_passphrase_optional": "Wachtwoordzin (optioneel)",
-  "mwallet_passphrase_title": "Apparaatwachtwoord (optioneel)",
   "mwallet_received": "Ontvangen",
   "mwallet_remove_server": "Server verwijderen",
   "mwallet_restore_branch": "Afleidingstak: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Besteedbaar",
   "mwallet_title": "Portemonnee",
   "mwallet_unlock_failed": "Ontgrendelen mislukt — verkeerde wachtwoordzin?",
-  "mwallet_verify_hint": "Voer de gevraagde woorden in vanuit uw geschreven back-up.",
-  "mwallet_verify_mismatch": "De woorden komen niet overeen met uw herstelzin.",
-  "mwallet_verify_title": "Bevestig uw back-up",
   "mwallet_import_bare_key": "Plak een volledige descriptor (wpkh(...) of tr(...)), geen kale sleutel.",
   "mwallet_import_bare_wif": "Dat is een kale privé-WIF-sleutel — omhul hem als wpkh(JOUW_WIF) om hem als portemonnee met één adres te importeren.",
   "mwallet_import_description": "Importeer een portemonnee vanuit zijn descriptors. Plak één of twee privé-descriptors — publieke (xpub) descriptors zijn alleen-lezen en worden nog niet ondersteund.",
