+++ web-wallet/src/assets/locales/pl.json
@@ -73,8 +73,7 @@
   "balance_available": "Dostepne saldo",
   "balance_history": "Historia salda",
   "base_target": "Cel podstawowy",
-  "bip39_passphrase": "Haslo BIP39",
-  "bip39_passphrase_info": "Mozesz dodac opcjonalne haslo (czasami nazywane 25. slowem) dla dodatkowego bezpieczenstwa. Tworzy to zupelnie inny portfel.",
+  "bip39_passphrase": "Hasło BIP39 (25. słowo)",
   "bip39_passphrase_warning": "Jesli ustawisz haslo, MUSISZ je pamietac. Bez hasla nie mozna odzyskac srodkow nawet z 24 slowami!",
   "bitcoin_pocx_network": "Siec Bitcoin PoCX",
   "block": "Blok",
@@ -137,7 +136,7 @@
   "compression": "Kompresja",
   "confirm": "Potwierdz",
   "confirm_backup_written": "Zapisalem fraze odzyskiwania i przechowalem ja bezpiecznie",
-  "confirm_bip39_passphrase": "Potwierdz haslo",
+  "confirm_bip39_passphrase": "Potwierdź 25. słowo",
   "confirm_send": "Potwierdz i wyslij",
   "confirm_send_message": "Sprawdz szczegoly transakcji przed wyslaniem.",
   "confirm_transaction": "Potwierdz transakcje",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Portfel utworzony. Każdy współsygnatariusz musi utworzyć swój portfel z tymi samymi kluczami, zanim otrzymacie środki.",
   "mwallet_address_failed": "Nie udało się pobrać adresu",
   "mwallet_add_to_contact": "Dodaj do kontaktów",
-  "mwallet_backup_ack": "Zapisałem 24 słowa",
-  "mwallet_backup_title": "Zapisz swoją frazę odzyskiwania",
-  "mwallet_backup_warning": "Te 24 słowa to jedyna kopia zapasowa portfela. Zapisz je na papierze i przechowuj offline. Każdy, kto je zna, może wydać Twoje środki.",
   "mwallet_branch_legacy": "starsza (coin type 0')",
   "mwallet_create_failed": "Nie udało się utworzyć portfela",
   "mwallet_create_wallet": "Utwórz nowy portfel",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Utwórz nowy portfel lub przywróć istniejący z frazy odzyskiwania.",
   "mwallet_onboarding_title": "Skonfiguruj swój portfel",
   "mwallet_passphrase_confirm": "Potwierdź hasło",
-  "mwallet_bip39_toggle": "Dodaj hasło BIP39 (25. słowo)",
-  "mwallet_bip39_label": "Hasło BIP39 (25. słowo)",
-  "mwallet_bip39_confirm": "Potwierdź 25. słowo",
-  "mwallet_bip39_warning": "To opcjonalne 25. słowo jest wplatane w Twoje klucze. Jeśli je ustawisz, SAMA fraza odzyskiwania NIE przywróci Twoich środków — musisz pamiętać także dokładnie to słowo. To nie jest hasło urządzenia powyżej.",
-  "mwallet_bip39_restore_warning": "Wpisz dokładnie to 25. słowo, z którym utworzono ten portfel. Błędne lub brakujące słowo wyprowadza inny, pusty portfel.",
+  "bip39_restore_warning": "Wpisz dokładnie to 25. słowo, z którym utworzono ten portfel. Błędne lub brakujące słowo wyprowadza inny, pusty portfel.",
   "mwallet_passphrase_hint": "Szyfruje seed zapisany na tym urządzeniu. Nie jest częścią frazy odzyskiwania — same słowa zawsze przywracają Twoje środki.",
   "mwallet_passphrase_label": "Hasło",
-  "mwallet_passphrase_mismatch": "Hasła nie są zgodne",
   "mwallet_passphrase_optional": "Hasło (opcjonalne)",
-  "mwallet_passphrase_title": "Hasło urządzenia (opcjonalne)",
   "mwallet_received": "Otrzymano",
   "mwallet_remove_server": "Usuń serwer",
   "mwallet_restore_branch": "Gałąź derywacji: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Do wydania",
   "mwallet_title": "Portfel",
   "mwallet_unlock_failed": "Nie udało się odblokować — złe hasło?",
-  "mwallet_verify_hint": "Wpisz wskazane słowa z zapisanej kopii.",
-  "mwallet_verify_mismatch": "Słowa nie pasują do frazy odzyskiwania.",
-  "mwallet_verify_title": "Potwierdź kopię zapasową",
   "mwallet_import_bare_key": "Wklej pełny deskryptor (wpkh(...) lub tr(...)), a nie sam klucz.",
   "mwallet_import_bare_wif": "To goły klucz prywatny WIF — opakuj go jako wpkh(TWÓJ_WIF), aby zaimportować go jako portfel z jednym adresem.",
   "mwallet_import_description": "Zaimportuj portfel z jego deskryptorów. Wklej jeden lub dwa prywatne deskryptory — publiczne deskryptory (xpub) są tylko do podglądu i nie są jeszcze obsługiwane.",
