+++ web-wallet/src/assets/locales/pt-br.json
@@ -73,8 +73,7 @@
   "balance_available": "Saldo disponível",
   "balance_history": "Histórico de saldo",
   "base_target": "Alvo Base",
-  "bip39_passphrase": "Frase de senha BIP39",
-  "bip39_passphrase_info": "Você pode adicionar uma frase de senha opcional (às vezes chamada de 25ª palavra) para segurança adicional. Isso cria uma carteira completamente diferente.",
+  "bip39_passphrase": "Frase de senha BIP39 (25ª palavra)",
   "bip39_passphrase_warning": "Se você definir uma frase de senha, DEVE lembrar dela. Sem a frase de senha, seus fundos não podem ser recuperados mesmo com as 24 palavras!",
   "bitcoin_pocx_network": "Rede Bitcoin PoCX",
   "block": "Bloco",
@@ -137,7 +136,7 @@
   "compression": "Compressão",
   "confirm": "Confirmar",
   "confirm_backup_written": "Eu escrevi minha frase de recuperação e a guardei com segurança",
-  "confirm_bip39_passphrase": "Confirmar frase de senha",
+  "confirm_bip39_passphrase": "Confirmar a 25ª palavra",
   "confirm_send": "Confirmar e enviar",
   "confirm_send_message": "Por favor, revise os detalhes da transação antes de enviar.",
   "confirm_transaction": "Confirmar transação",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Carteira criada. Cada cossignatário deve criar sua carteira com as mesmas chaves antes de receber fundos.",
   "mwallet_address_failed": "Falha ao obter um endereço",
   "mwallet_add_to_contact": "Adicionar aos contatos",
-  "mwallet_backup_ack": "Anotei as 24 palavras",
-  "mwallet_backup_title": "Anote sua frase de recuperação",
-  "mwallet_backup_warning": "Estas 24 palavras são o único backup da sua carteira. Anote-as em papel e guarde-as offline. Qualquer pessoa que as conheça pode gastar seus fundos.",
   "mwallet_branch_legacy": "legado (coin type 0')",
   "mwallet_create_failed": "Falha ao criar a carteira",
   "mwallet_create_wallet": "Criar nova carteira",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Crie uma nova carteira ou restaure uma existente a partir da frase de recuperação.",
   "mwallet_onboarding_title": "Configure sua carteira",
   "mwallet_passphrase_confirm": "Confirmar senha",
-  "mwallet_bip39_toggle": "Adicionar frase de senha BIP39 (25ª palavra)",
-  "mwallet_bip39_label": "Frase de senha BIP39 (25ª palavra)",
-  "mwallet_bip39_confirm": "Confirmar a 25ª palavra",
-  "mwallet_bip39_warning": "Esta 25ª palavra opcional é incorporada às suas chaves. Se você a definir, a frase de recuperação SOZINHA NÃO restaurará seus fundos — você também precisa lembrar exatamente esta palavra. Não é a frase de senha do dispositivo acima.",
-  "mwallet_bip39_restore_warning": "Digite exatamente a 25ª palavra com a qual esta carteira foi criada. Uma palavra errada ou ausente deriva uma carteira diferente e vazia.",
+  "bip39_restore_warning": "Digite exatamente a 25ª palavra com a qual esta carteira foi criada. Uma palavra errada ou ausente deriva uma carteira diferente e vazia.",
   "mwallet_passphrase_hint": "Criptografa a semente armazenada neste dispositivo. Não faz parte da frase de recuperação — as palavras sozinhas sempre restauram seus fundos.",
   "mwallet_passphrase_label": "Senha",
-  "mwallet_passphrase_mismatch": "As senhas não correspondem",
   "mwallet_passphrase_optional": "Senha (opcional)",
-  "mwallet_passphrase_title": "Senha do dispositivo (opcional)",
   "mwallet_received": "Recebido",
   "mwallet_remove_server": "Remover servidor",
   "mwallet_restore_branch": "Ramo de derivação: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Disponível",
   "mwallet_title": "Carteira",
   "mwallet_unlock_failed": "Não foi possível desbloquear — senha incorreta?",
-  "mwallet_verify_hint": "Digite as palavras solicitadas do seu backup escrito.",
-  "mwallet_verify_mismatch": "As palavras não correspondem à sua frase de recuperação.",
-  "mwallet_verify_title": "Confirme seu backup",
   "mwallet_import_bare_key": "Cole um descritor completo (wpkh(...) ou tr(...)), não uma chave avulsa.",
   "mwallet_import_bare_wif": "Isso é uma chave privada WIF avulsa — envolva-a como wpkh(SUA_WIF) para importá-la como carteira de endereço único.",
   "mwallet_import_description": "Importe uma carteira a partir de seus descritores. Cole um ou dois descritores privados — descritores públicos (xpub) são somente observação e ainda não são suportados.",
