+++ web-wallet/src/assets/locales/ro.json
@@ -73,8 +73,7 @@
   "balance_available": "Sold disponibil",
   "balance_history": "Istoric sold",
   "base_target": "Țintă de bază",
-  "bip39_passphrase": "Parolă BIP39",
-  "bip39_passphrase_info": "Puteți adăuga o parolă opțională (uneori numită al 25-lea cuvânt) pentru securitate suplimentară. Aceasta creează un portofel complet diferit.",
+  "bip39_passphrase": "Parolă BIP39 (al 25-lea cuvânt)",
   "bip39_passphrase_warning": "Dacă setați o parolă, TREBUIE să o țineți minte. Fără parolă, fondurile nu pot fi recuperate nici măcar cu cele 24 de cuvinte!",
   "bitcoin_pocx_network": "Rețea Bitcoin PoCX",
   "block": "Bloc",
@@ -137,7 +136,7 @@
   "compression": "Compresie",
   "confirm": "Confirmă",
   "confirm_backup_written": "Am scris fraza de recuperare și am păstrat-o în siguranță",
-  "confirm_bip39_passphrase": "Confirmă parola",
+  "confirm_bip39_passphrase": "Confirmă al 25-lea cuvânt",
   "confirm_send": "Confirmă și trimite",
   "confirm_send_message": "Vă rugăm să verificați detaliile tranzacției înainte de a trimite.",
   "confirm_transaction": "Confirmă tranzacția",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Portofel creat. Fiecare co-semnatar trebuie să își creeze portofelul cu aceleași chei înainte de a primi fonduri.",
   "mwallet_address_failed": "Nu s-a putut obține o adresă",
   "mwallet_add_to_contact": "Adaugă la contacte",
-  "mwallet_backup_ack": "Am notat cele 24 de cuvinte",
-  "mwallet_backup_title": "Notați fraza de recuperare",
-  "mwallet_backup_warning": "Aceste 24 de cuvinte sunt singura copie de rezervă a portofelului. Notați-le pe hârtie și păstrați-le offline. Oricine le cunoaște vă poate cheltui fondurile.",
   "mwallet_branch_legacy": "veche (coin type 0')",
   "mwallet_create_failed": "Portofelul nu a putut fi creat",
   "mwallet_create_wallet": "Creează portofel nou",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Creați un portofel nou sau restaurați unul existent din fraza de recuperare.",
   "mwallet_onboarding_title": "Configurați-vă portofelul",
   "mwallet_passphrase_confirm": "Confirmați parola",
-  "mwallet_bip39_toggle": "Adaugă parolă BIP39 (al 25-lea cuvânt)",
-  "mwallet_bip39_label": "Parolă BIP39 (al 25-lea cuvânt)",
-  "mwallet_bip39_confirm": "Confirmă al 25-lea cuvânt",
-  "mwallet_bip39_warning": "Acest al 25-lea cuvânt opțional este încorporat în cheile tale. Dacă îl setezi, fraza de recuperare SINGURĂ NU îți va restaura fondurile — trebuie să reții și acest cuvânt exact. Nu este parola dispozitivului de mai sus.",
-  "mwallet_bip39_restore_warning": "Introdu exact al 25-lea cuvânt cu care a fost creat acest portofel. Un cuvânt greșit sau lipsă derivă un portofel diferit, gol.",
+  "bip39_restore_warning": "Introdu exact al 25-lea cuvânt cu care a fost creat acest portofel. Un cuvânt greșit sau lipsă derivă un portofel diferit, gol.",
   "mwallet_passphrase_hint": "Criptează seed-ul stocat pe acest dispozitiv. Nu face parte din fraza de recuperare — cuvintele singure vă restaurează întotdeauna fondurile.",
   "mwallet_passphrase_label": "Parolă",
-  "mwallet_passphrase_mismatch": "Parolele nu corespund",
   "mwallet_passphrase_optional": "Parolă (opțional)",
-  "mwallet_passphrase_title": "Parola dispozitivului (opțional)",
   "mwallet_received": "Primit",
   "mwallet_remove_server": "Elimină serverul",
   "mwallet_restore_branch": "Ramura de derivare: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Disponibil",
   "mwallet_title": "Portofel",
   "mwallet_unlock_failed": "Deblocarea a eșuat — parolă greșită?",
-  "mwallet_verify_hint": "Introduceți cuvintele solicitate din copia scrisă.",
-  "mwallet_verify_mismatch": "Cuvintele nu corespund frazei de recuperare.",
-  "mwallet_verify_title": "Confirmați copia de rezervă",
   "mwallet_import_bare_key": "Lipiți un descriptor complet (wpkh(...) sau tr(...)), nu doar o cheie.",
   "mwallet_import_bare_wif": "Aceasta este o cheie privată WIF simplă — înfășurați-o ca wpkh(WIF_UL_DVS) pentru a o importa ca portofel cu o singură adresă.",
   "mwallet_import_description": "Importați un portofel din descriptorii săi. Lipiți unul sau doi descriptori privați — descriptorii publici (xpub) sunt doar de observare și nu sunt încă suportați.",
