+++ web-wallet/src/assets/locales/ru.json
@@ -73,8 +73,7 @@
   "balance_available": "Доступный баланс",
   "balance_history": "История баланса",
   "base_target": "Базовая цель",
-  "bip39_passphrase": "BIP39 парольная фраза",
-  "bip39_passphrase_info": "Вы можете добавить опциональную парольную фразу (иногда называемую 25-м словом) для дополнительной безопасности. Это создаст совершенно другой кошелек.",
+  "bip39_passphrase": "Парольная фраза BIP39 (25-е слово)",
   "bip39_passphrase_warning": "Если вы установите парольную фразу, вы ДОЛЖНЫ её запомнить. Без парольной фразы ваши средства не могут быть восстановлены даже с 24 словами!",
   "bitcoin_pocx_network": "Сеть Bitcoin PoCX",
   "block": "Блок",
@@ -137,7 +136,7 @@
   "compression": "Сжатие",
   "confirm": "Подтвердить",
   "confirm_backup_written": "Я записал фразу восстановления и сохранил её в безопасном месте",
-  "confirm_bip39_passphrase": "Подтвердите парольную фразу",
+  "confirm_bip39_passphrase": "Подтвердите 25-е слово",
   "confirm_send": "Подтвердить и отправить",
   "confirm_send_message": "Пожалуйста, проверьте детали транзакции перед отправкой.",
   "confirm_transaction": "Подтвердить транзакцию",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Кошелёк создан. Каждый соподписант должен создать свой кошелёк с теми же ключами до получения средств.",
   "mwallet_address_failed": "Не удалось получить адрес",
   "mwallet_add_to_contact": "Добавить в контакты",
-  "mwallet_backup_ack": "Я записал 24 слова",
-  "mwallet_backup_title": "Запишите вашу фразу восстановления",
-  "mwallet_backup_warning": "Эти 24 слова — единственная резервная копия вашего кошелька. Запишите их на бумаге и храните офлайн. Любой, кто их знает, может потратить ваши средства.",
   "mwallet_branch_legacy": "устаревшая (coin type 0')",
   "mwallet_create_failed": "Не удалось создать кошелек",
   "mwallet_create_wallet": "Создать новый кошелек",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Создайте новый кошелек или восстановите существующий из фразы восстановления.",
   "mwallet_onboarding_title": "Настройте свой кошелек",
   "mwallet_passphrase_confirm": "Подтвердите пароль",
-  "mwallet_bip39_toggle": "Добавить парольную фразу BIP39 (25-е слово)",
-  "mwallet_bip39_label": "Парольная фраза BIP39 (25-е слово)",
-  "mwallet_bip39_confirm": "Подтвердите 25-е слово",
-  "mwallet_bip39_warning": "Это необязательное 25-е слово встраивается в ваши ключи. Если вы его зададите, фраза восстановления САМА ПО СЕБЕ НЕ восстановит ваши средства — нужно помнить и именно это слово. Это не парольная фраза устройства выше.",
-  "mwallet_bip39_restore_warning": "Введите точно то 25-е слово, с которым был создан этот кошелёк. Неверное или пропущенное слово выводит другой, пустой кошелёк.",
+  "bip39_restore_warning": "Введите точно то 25-е слово, с которым был создан этот кошелёк. Неверное или пропущенное слово выводит другой, пустой кошелёк.",
   "mwallet_passphrase_hint": "Шифрует seed, хранящийся на этом устройстве. Он не является частью фразы восстановления — сами слова всегда восстанавливают ваши средства.",
   "mwallet_passphrase_label": "Пароль",
-  "mwallet_passphrase_mismatch": "Пароли не совпадают",
   "mwallet_passphrase_optional": "Пароль (необязательно)",
-  "mwallet_passphrase_title": "Пароль устройства (необязательно)",
   "mwallet_received": "Получено",
   "mwallet_remove_server": "Удалить сервер",
   "mwallet_restore_branch": "Ветвь деривации: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Доступно",
   "mwallet_title": "Кошелек",
   "mwallet_unlock_failed": "Не удалось разблокировать — неверный пароль?",
-  "mwallet_verify_hint": "Введите запрошенные слова из вашей письменной копии.",
-  "mwallet_verify_mismatch": "Слова не совпадают с вашей фразой восстановления.",
-  "mwallet_verify_title": "Подтвердите резервную копию",
   "mwallet_import_bare_key": "Вставьте полный дескриптор (wpkh(...) или tr(...)), а не просто ключ.",
   "mwallet_import_bare_wif": "Это голый приватный ключ WIF — оберните его как wpkh(ВАШ_WIF), чтобы импортировать его как кошелек с одним адресом.",
   "mwallet_import_description": "Импортируйте кошелек из его дескрипторов. Вставьте один или два приватных дескриптора — публичные дескрипторы (xpub) предназначены только для наблюдения и пока не поддерживаются.",
