+++ web-wallet/src/assets/locales/sk.json
@@ -73,8 +73,7 @@
   "balance_available": "Dostupny zostatok",
   "balance_history": "Historia zostatku",
   "base_target": "Zakladny ciel",
-  "bip39_passphrase": "BIP39 heslo",
-  "bip39_passphrase_info": "Mozete pridat volitelne heslo (niekedy nazyvane 25. slovo) pre dodatocne zabezpecenie. Vytvori to uplne inu penazenku.",
+  "bip39_passphrase": "BIP39 heslo (25. slovo)",
   "bip39_passphrase_warning": "Ak nastavite heslo, MUSITE si ho pamatat. Bez hesla sa prostriedky nedaju obnovit ani s 24 slovami!",
   "bitcoin_pocx_network": "Siet Bitcoin PoCX",
   "block": "Blok",
@@ -137,7 +136,7 @@
   "compression": "Kompresia",
   "confirm": "Potvrdit",
   "confirm_backup_written": "Zapisal som si svoju obnovovaciu frazu a bezpecne ju ulozil",
-  "confirm_bip39_passphrase": "Potvrdit heslo",
+  "confirm_bip39_passphrase": "Potvrdiť 25. slovo",
   "confirm_send": "Potvrdit a odoslat",
   "confirm_send_message": "Skontrolujte podrobnosti transakcie pred odoslanim.",
   "confirm_transaction": "Potvrdit transakciu",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Peňaženka vytvorená. Každý spolupodpisovateľ musí pred prijímaním prostriedkov vytvoriť svoju peňaženku s rovnakými kľúčmi.",
   "mwallet_address_failed": "Nepodarilo sa získať adresu",
   "mwallet_add_to_contact": "Pridať do kontaktov",
-  "mwallet_backup_ack": "Zapísal som si 24 slov",
-  "mwallet_backup_title": "Zapíšte si svoju obnovovaciu frázu",
-  "mwallet_backup_warning": "Týchto 24 slov je jedinou zálohou vašej peňaženky. Zapíšte si ich na papier a uchovávajte offline. Ktokoľvek ich pozná, môže minúť vaše prostriedky.",
   "mwallet_branch_legacy": "staršia (coin type 0')",
   "mwallet_create_failed": "Peňaženku sa nepodarilo vytvoriť",
   "mwallet_create_wallet": "Vytvoriť novú peňaženku",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Vytvorte novú peňaženku alebo obnovte existujúcu z obnovovacej frázy.",
   "mwallet_onboarding_title": "Nastavte si peňaženku",
   "mwallet_passphrase_confirm": "Potvrdiť heslo",
-  "mwallet_bip39_toggle": "Pridať BIP39 heslo (25. slovo)",
-  "mwallet_bip39_label": "BIP39 heslo (25. slovo)",
-  "mwallet_bip39_confirm": "Potvrdiť 25. slovo",
-  "mwallet_bip39_warning": "Toto voliteľné 25. slovo sa stáva súčasťou vašich kľúčov. Ak ho nastavíte, obnovovacia fráza SAMA vaše prostriedky NEOBNOVÍ — musíte si pamätať aj presne toto slovo. Nie je to heslo zariadenia vyššie.",
-  "mwallet_bip39_restore_warning": "Zadajte presné 25. slovo, s ktorým bola táto peňaženka vytvorená. Nesprávne alebo chýbajúce slovo odvodí inú, prázdnu peňaženku.",
+  "bip39_restore_warning": "Zadajte presné 25. slovo, s ktorým bola táto peňaženka vytvorená. Nesprávne alebo chýbajúce slovo odvodí inú, prázdnu peňaženku.",
   "mwallet_passphrase_hint": "Šifruje seed uložený na tomto zariadení. Nie je súčasťou obnovovacej frázy — samotné slová vždy obnovia vaše prostriedky.",
   "mwallet_passphrase_label": "Heslo",
-  "mwallet_passphrase_mismatch": "Heslá sa nezhodujú",
   "mwallet_passphrase_optional": "Heslo (voliteľné)",
-  "mwallet_passphrase_title": "Heslo zariadenia (voliteľné)",
   "mwallet_received": "Prijaté",
   "mwallet_remove_server": "Odstrániť server",
   "mwallet_restore_branch": "Derivačná vetva: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "K dispozícii",
   "mwallet_title": "Peňaženka",
   "mwallet_unlock_failed": "Odomknutie zlyhalo — nesprávne heslo?",
-  "mwallet_verify_hint": "Zadajte požadované slová zo svojej písomnej zálohy.",
-  "mwallet_verify_mismatch": "Slová sa nezhodujú s vašou obnovovacou frázou.",
-  "mwallet_verify_title": "Potvrďte svoju zálohu",
   "mwallet_import_bare_key": "Vložte úplný deskriptor (wpkh(...) alebo tr(...)), nie samotný kľúč.",
   "mwallet_import_bare_wif": "To je samotný súkromný kľúč WIF — obaľte ho ako wpkh(VÁŠ_WIF) a naimportujte ho ako peňaženku s jednou adresou.",
   "mwallet_import_description": "Importujte peňaženku z jej deskriptorov. Vložte jeden alebo dva súkromné deskriptory — verejné (xpub) deskriptory sú len na sledovanie a zatiaľ nie sú podporované.",
