+++ web-wallet/src/assets/locales/sr.json
@@ -73,8 +73,7 @@
   "balance_available": "Доступно стање",
   "balance_history": "Историја стања",
   "base_target": "Основна мета",
-  "bip39_passphrase": "BIP39 лозинка",
-  "bip39_passphrase_info": "Можете додати опциону лозинку (понекад звану 25. реч) за додатну сигурност. Ово ствара потпуно другачији новчаник.",
+  "bip39_passphrase": "BIP39 лозинка (25. реч)",
   "bip39_passphrase_warning": "Ако поставите лозинку, МОРАТЕ је запамтити. Без лозинке ваша средства не могу бити опорављена чак ни са 24 речи!",
   "bitcoin_pocx_network": "Bitcoin PoCX мрежа",
   "block": "Блок",
@@ -137,7 +136,7 @@
   "compression": "Компресија",
   "confirm": "Потврди",
   "confirm_backup_written": "Записао сам фразу за опоравак и сачувао је на сигурном месту",
-  "confirm_bip39_passphrase": "Потврдите лозинку",
+  "confirm_bip39_passphrase": "Потврди 25. реч",
   "confirm_send": "Потврди и пошаљи",
   "confirm_send_message": "Молимо прегледајте детаље трансакције пре слања.",
   "confirm_transaction": "Потврдите трансакцију",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Новчаник је креиран. Сваки супотписник мора креирати свој новчаник са истим кључевима пре примања средстава.",
   "mwallet_address_failed": "Није могуће добити адресу",
   "mwallet_add_to_contact": "Додај у контакте",
-  "mwallet_backup_ack": "Записао сам 24 речи",
-  "mwallet_backup_title": "Запишите своју фразу за опоравак",
-  "mwallet_backup_warning": "Ове 24 речи су једина резервна копија вашег новчаника. Запишите их на папир и чувајте ван мреже. Свако ко их зна може потрошити ваша средства.",
   "mwallet_branch_legacy": "наслеђена (coin type 0')",
   "mwallet_create_failed": "Новчаник није могуће креирати",
   "mwallet_create_wallet": "Креирај нови новчаник",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Креирајте нови новчаник или вратите постојећи из фразе за опоравак.",
   "mwallet_onboarding_title": "Подесите свој новчаник",
   "mwallet_passphrase_confirm": "Потврди лозинку",
-  "mwallet_bip39_toggle": "Додај BIP39 лозинку (25. реч)",
-  "mwallet_bip39_label": "BIP39 лозинка (25. реч)",
-  "mwallet_bip39_confirm": "Потврди 25. реч",
-  "mwallet_bip39_warning": "Ова опциона 25. реч се уграђује у ваше кључеве. Ако је поставите, фраза за опоравак САМА НЕЋЕ вратити ваша средства — морате запамтити и тачно ову реч. То није лозинка уређаја изнад.",
-  "mwallet_bip39_restore_warning": "Унесите тачну 25. реч са којом је овај новчаник направљен. Погрешна или изостављена реч изводи други, празан новчаник.",
+  "bip39_restore_warning": "Унесите тачну 25. реч са којом је овај новчаник направљен. Погрешна или изостављена реч изводи други, празан новчаник.",
   "mwallet_passphrase_hint": "Шифрује seed сачуван на овом уређају. Није део фразе за опоравак — саме речи увек враћају ваша средства.",
   "mwallet_passphrase_label": "Лозинка",
-  "mwallet_passphrase_mismatch": "Лозинке се не подударају",
   "mwallet_passphrase_optional": "Лозинка (опционо)",
-  "mwallet_passphrase_title": "Лозинка уређаја (опционо)",
   "mwallet_received": "Примљено",
   "mwallet_remove_server": "Уклони сервер",
   "mwallet_restore_branch": "Грана деривације: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Расположиво",
   "mwallet_title": "Новчаник",
   "mwallet_unlock_failed": "Откључавање није успело — погрешна лозинка?",
-  "mwallet_verify_hint": "Унесите тражене речи из записане копије.",
-  "mwallet_verify_mismatch": "Речи се не подударају са вашом фразом за опоравак.",
-  "mwallet_verify_title": "Потврдите резервну копију",
   "mwallet_import_bare_key": "Налепите потпун дескриптор (wpkh(...) или tr(...)), а не сам кључ.",
   "mwallet_import_bare_wif": "То је го приватни WIF кључ — умотајте га као wpkh(ВАШ_WIF) да бисте га увезли као новчаник са једном адресом.",
   "mwallet_import_description": "Увезите новчаник из његових дескриптора. Налепите један или два приватна дескриптора — јавни (xpub) дескриптори су само за посматрање и још нису подржани.",
