+++ web-wallet/src/assets/locales/tr.json
@@ -73,8 +73,7 @@
   "balance_available": "Kullanılabilir Bakiye",
   "balance_history": "Bakiye Geçmişi",
   "base_target": "Temel Hedef",
-  "bip39_passphrase": "BIP39 Parolası",
-  "bip39_passphrase_info": "Ek güvenlik için isteğe bağlı bir parola (bazen 25. kelime olarak adlandırılır) ekleyebilirsiniz. Bu tamamen farklı bir cüzdan oluşturur.",
+  "bip39_passphrase": "BIP39 parolası (25. kelime)",
   "bip39_passphrase_warning": "Parola belirlerseniz, MUTLAKA hatırlamalısınız. Parola olmadan, 24 kelimeyle bile fonlarınız kurtarılamaz!",
   "bitcoin_pocx_network": "Bitcoin PoCX Ağı",
   "block": "Blok",
@@ -137,7 +136,7 @@
   "compression": "Sıkıştırma",
   "confirm": "Onayla",
   "confirm_backup_written": "Kurtarma ifademi yazdım ve güvenli bir şekilde sakladım",
-  "confirm_bip39_passphrase": "Parolayı Onayla",
+  "confirm_bip39_passphrase": "25. kelimeyi onayla",
   "confirm_send": "Onayla ve Gönder",
   "confirm_send_message": "Göndermeden önce işlem detaylarını inceleyin.",
   "confirm_transaction": "İşlemi Onayla",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Cüzdan oluşturuldu. Fon almadan önce her ortak imzalayan aynı anahtarlarla kendi cüzdanını oluşturmalıdır.",
   "mwallet_address_failed": "Adres alınamadı",
   "mwallet_add_to_contact": "Kişilere ekle",
-  "mwallet_backup_ack": "24 kelimeyi yazdım",
-  "mwallet_backup_title": "Kurtarma ifadenizi yazın",
-  "mwallet_backup_warning": "Bu 24 kelime cüzdanınızın tek yedeğidir. Kağıda yazın ve çevrimdışı saklayın. Bunları bilen herkes fonlarınızı harcayabilir.",
   "mwallet_branch_legacy": "eski (coin type 0')",
   "mwallet_create_failed": "Cüzdan oluşturulamadı",
   "mwallet_create_wallet": "Yeni cüzdan oluştur",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Yeni bir cüzdan oluşturun veya mevcut bir cüzdanı kurtarma ifadesinden geri yükleyin.",
   "mwallet_onboarding_title": "Cüzdanınızı kurun",
   "mwallet_passphrase_confirm": "Parolayı onayla",
-  "mwallet_bip39_toggle": "BIP39 parolası ekle (25. kelime)",
-  "mwallet_bip39_label": "BIP39 parolası (25. kelime)",
-  "mwallet_bip39_confirm": "25. kelimeyi onayla",
-  "mwallet_bip39_warning": "Bu isteğe bağlı 25. kelime anahtarlarınıza işlenir. Ayarlarsanız, kurtarma ifadeniz TEK BAŞINA fonlarınızı GERİ GETİRMEZ — bu kelimeyi de tam olarak hatırlamanız gerekir. Yukarıdaki cihaz parolası değildir.",
-  "mwallet_bip39_restore_warning": "Bu cüzdanın oluşturulduğu 25. kelimeyi tam olarak girin. Yanlış veya eksik kelime farklı, boş bir cüzdan türetir.",
+  "bip39_restore_warning": "Bu cüzdanın oluşturulduğu 25. kelimeyi tam olarak girin. Yanlış veya eksik kelime farklı, boş bir cüzdan türetir.",
   "mwallet_passphrase_hint": "Bu cihazda saklanan seed'i şifreler. Kurtarma ifadesinin parçası değildir — kelimeler tek başına fonlarınızı her zaman geri yükler.",
   "mwallet_passphrase_label": "Parola",
-  "mwallet_passphrase_mismatch": "Parolalar eşleşmiyor",
   "mwallet_passphrase_optional": "Parola (isteğe bağlı)",
-  "mwallet_passphrase_title": "Cihaz parolası (isteğe bağlı)",
   "mwallet_received": "Alındı",
   "mwallet_remove_server": "Sunucuyu kaldır",
   "mwallet_restore_branch": "Türetme dalı: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Harcanabilir",
   "mwallet_title": "Cüzdan",
   "mwallet_unlock_failed": "Kilit açılamadı — yanlış parola?",
-  "mwallet_verify_hint": "Yazılı yedeğinizden istenen kelimeleri girin.",
-  "mwallet_verify_mismatch": "Kelimeler kurtarma ifadenizle eşleşmiyor.",
-  "mwallet_verify_title": "Yedeğinizi onaylayın",
   "mwallet_import_bare_key": "Tam bir tanımlayıcı yapıştırın (wpkh(...) veya tr(...)), yalın bir anahtar değil.",
   "mwallet_import_bare_wif": "Bu çıplak bir özel WIF anahtarı — tek adresli bir cüzdan olarak içe aktarmak için wpkh(WIF_ANAHTARINIZ) biçiminde sarın.",
   "mwallet_import_description": "Bir cüzdanı tanımlayıcılarından içe aktarın. Bir veya iki özel tanımlayıcı yapıştırın — genel (xpub) tanımlayıcılar yalnızca izleme içindir ve henüz desteklenmiyor.",
