+++ web-wallet/src/assets/locales/uk.json
@@ -73,8 +73,7 @@
   "balance_available": "Доступний залишок",
   "balance_history": "Історія балансу",
   "base_target": "Базова ціль",
-  "bip39_passphrase": "BIP39 парольна фраза",
-  "bip39_passphrase_info": "Ви можете додати опціональну парольну фразу (іноді звану 25-м словом) для додаткової безпеки. Це створить зовсім інший гаманець.",
+  "bip39_passphrase": "Парольна фраза BIP39 (25-те слово)",
   "bip39_passphrase_warning": "Якщо ви встановите парольну фразу, ви ПОВИННІ її запам'ятати. Без парольної фрази ваші кошти не можуть бути відновлені навіть з 24 словами!",
   "bitcoin_pocx_network": "Мережа Bitcoin PoCX",
   "block": "Блок",
@@ -137,7 +136,7 @@
   "compression": "Стиснення",
   "confirm": "Підтвердити",
   "confirm_backup_written": "Я записав фразу відновлення та зберіг її в безпечному місці",
-  "confirm_bip39_passphrase": "Підтвердіть парольну фразу",
+  "confirm_bip39_passphrase": "Підтвердьте 25-те слово",
   "confirm_send": "Підтвердити та надіслати",
   "confirm_send_message": "Будь ласка, перевірте деталі транзакції перед надсиланням.",
   "confirm_transaction": "Підтвердити транзакцію",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "Гаманець створено. Кожен співпідписант має створити свій гаманець з тими самими ключами перед отриманням коштів.",
   "mwallet_address_failed": "Не вдалося отримати адресу",
   "mwallet_add_to_contact": "Додати до контактів",
-  "mwallet_backup_ack": "Я записав 24 слова",
-  "mwallet_backup_title": "Запишіть вашу фразу відновлення",
-  "mwallet_backup_warning": "Ці 24 слова — єдина резервна копія вашого гаманця. Запишіть їх на папері та зберігайте офлайн. Будь-хто, хто їх знає, може витратити ваші кошти.",
   "mwallet_branch_legacy": "застаріла (coin type 0')",
   "mwallet_create_failed": "Не вдалося створити гаманець",
   "mwallet_create_wallet": "Створити новий гаманець",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "Створіть новий гаманець або відновіть наявний із фрази відновлення.",
   "mwallet_onboarding_title": "Налаштуйте свій гаманець",
   "mwallet_passphrase_confirm": "Підтвердьте пароль",
-  "mwallet_bip39_toggle": "Додати парольну фразу BIP39 (25-те слово)",
-  "mwallet_bip39_label": "Парольна фраза BIP39 (25-те слово)",
-  "mwallet_bip39_confirm": "Підтвердьте 25-те слово",
-  "mwallet_bip39_warning": "Це необов'язкове 25-те слово вбудовується у ваші ключі. Якщо ви його задасте, фраза відновлення САМА ПО СОБІ НЕ відновить ваші кошти — потрібно пам'ятати й саме це слово. Це не парольна фраза пристрою вище.",
-  "mwallet_bip39_restore_warning": "Введіть точно те 25-те слово, з яким було створено цей гаманець. Неправильне або пропущене слово виводить інший, порожній гаманець.",
+  "bip39_restore_warning": "Введіть точно те 25-те слово, з яким було створено цей гаманець. Неправильне або пропущене слово виводить інший, порожній гаманець.",
   "mwallet_passphrase_hint": "Шифрує seed, збережений на цьому пристрої. Він не є частиною фрази відновлення — самі слова завжди відновлюють ваші кошти.",
   "mwallet_passphrase_label": "Пароль",
-  "mwallet_passphrase_mismatch": "Паролі не збігаються",
   "mwallet_passphrase_optional": "Пароль (необов'язково)",
-  "mwallet_passphrase_title": "Пароль пристрою (необов'язково)",
   "mwallet_received": "Отримано",
   "mwallet_remove_server": "Видалити сервер",
   "mwallet_restore_branch": "Гілка деривації: {branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "Доступно",
   "mwallet_title": "Гаманець",
   "mwallet_unlock_failed": "Не вдалося розблокувати — неправильний пароль?",
-  "mwallet_verify_hint": "Введіть запитані слова з вашої письмової копії.",
-  "mwallet_verify_mismatch": "Слова не збігаються з вашою фразою відновлення.",
-  "mwallet_verify_title": "Підтвердьте резервну копію",
   "mwallet_import_bare_key": "Вставте повний дескриптор (wpkh(...) або tr(...)), а не сам ключ.",
   "mwallet_import_bare_wif": "Це голий приватний ключ WIF — оберніть його як wpkh(ВАШ_WIF), щоб імпортувати його як гаманець з однією адресою.",
   "mwallet_import_description": "Імпортуйте гаманець із його дескрипторів. Вставте один або два приватні дескриптори — публічні дескриптори (xpub) призначені лише для перегляду й поки не підтримуються.",
