+++ web-wallet/src/assets/locales/zh-cn.json
@@ -73,8 +73,7 @@
   "balance_available": "可用余额",
   "balance_history": "余额历史",
   "base_target": "基本目标",
-  "bip39_passphrase": "BIP39 密码短语",
-  "bip39_passphrase_info": "您可以添加可选的密码短语（有时称为第25个词）以增强安全性。这将创建一个完全不同的钱包。",
+  "bip39_passphrase": "BIP39 密码短语（第25个词）",
   "bip39_passphrase_warning": "如果您设置了密码短语，必须牢记它。没有密码短语，即使有24个单词也无法恢复您的资金！",
   "bitcoin_pocx_network": "Bitcoin PoCX 网络",
   "block": "区块",
@@ -137,7 +136,7 @@
   "compression": "压缩",
   "confirm": "确认",
   "confirm_backup_written": "我已写下恢复短语并妥善保管",
-  "confirm_bip39_passphrase": "确认密码短语",
+  "confirm_bip39_passphrase": "确认第25个词",
   "confirm_send": "确认并发送",
   "confirm_send_message": "发送前请检查交易详情。",
   "confirm_transaction": "确认交易",
@@ -1079,9 +1078,6 @@
   "msig_created_hint": "钱包已创建。每位共同签名者必须使用相同的密钥创建其钱包后，方可接收资金。",
   "mwallet_address_failed": "获取地址失败",
   "mwallet_add_to_contact": "添加到联系人",
-  "mwallet_backup_ack": "我已记下这24个单词",
-  "mwallet_backup_title": "记下您的恢复短语",
-  "mwallet_backup_warning": "这24个单词是您钱包的唯一备份。请写在纸上并离线保存。任何知道它们的人都可以花费您的资金。",
   "mwallet_branch_legacy": "旧版（coin type 0'）",
   "mwallet_create_failed": "创建钱包失败",
   "mwallet_create_wallet": "创建新钱包",
@@ -1099,16 +1095,10 @@
   "mwallet_onboarding_intro": "创建新钱包，或从恢复短语恢复现有钱包。",
   "mwallet_onboarding_title": "设置您的钱包",
   "mwallet_passphrase_confirm": "确认口令",
-  "mwallet_bip39_toggle": "添加 BIP39 密码短语（第25个词）",
-  "mwallet_bip39_label": "BIP39 密码短语（第25个词）",
-  "mwallet_bip39_confirm": "确认第25个词",
-  "mwallet_bip39_warning": "这个可选的第25个词会融入您的密钥。一旦设置，仅凭恢复短语将无法恢复您的资金 — 您还必须准确记住这个词。它不是上方的设备密码短语。",
-  "mwallet_bip39_restore_warning": "请输入创建此钱包时使用的第25个词。词语错误或缺失会派生出另一个空钱包。",
+  "bip39_restore_warning": "请输入创建此钱包时使用的第25个词。词语错误或缺失会派生出另一个空钱包。",
   "mwallet_passphrase_hint": "加密存储在此设备上的种子。它不是恢复短语的一部分 — 仅凭这些单词就能随时恢复您的资金。",
   "mwallet_passphrase_label": "口令",
-  "mwallet_passphrase_mismatch": "口令不匹配",
   "mwallet_passphrase_optional": "口令（可选）",
-  "mwallet_passphrase_title": "设备口令（可选）",
   "mwallet_received": "已接收",
   "mwallet_remove_server": "移除服务器",
   "mwallet_restore_branch": "派生分支：{branch}",
@@ -1134,9 +1124,6 @@
   "mwallet_spendable": "可用",
   "mwallet_title": "钱包",
   "mwallet_unlock_failed": "无法解锁 — 口令错误？",
-  "mwallet_verify_hint": "从您记下的备份中输入指定的单词。",
-  "mwallet_verify_mismatch": "单词与您的恢复短语不匹配。",
-  "mwallet_verify_title": "确认您的备份",
   "mwallet_import_bare_key": "请粘贴完整的描述符（wpkh(...) 或 tr(...)），而不是单独的密钥。",
   "mwallet_import_bare_wif": "这是一个裸 WIF 私钥 — 将其包装为 wpkh(你的WIF) 即可作为单地址钱包导入。",
   "mwallet_import_description": "从描述符导入钱包。粘贴一个或两个私有描述符 — 公开（xpub）描述符仅用于观察，暂不支持。",
